% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accession_gene_table}
\alias{accession_gene_table}
\title{UniProt human protein accession number to gene name mapping}
\format{
A tibble object with 167,966 rows and 3 columns:
\describe{
  \item{accession_number}{<chr> UniProt protein accession number (e.g. P10415)}
  \item{gene}{<chr> gene name (e.g. BCL2), first gene in all_genes if multiple}
  \item{all_genes}{<chr> comma-delimited gene names mapped to accession_number}
}
}
\source{
Human ID mapping file (2023-09-13 release) downloaded from UniProt:  
\url{https://www.uniprot.org}
}
\usage{
accession_gene_table
}
\description{
A dataset containing 167,966 UniProt accession number and gene name pairings.
}
\references{
UniProt Consortium. UniProt: the Universal Protein Knowledgebase in 2023.
Nucleic Acids Res. 2023;51(D1):D523-D531. doi:10.1093/nar/gkac1052
}
\seealso{
Other datasets: 
\code{\link{bioplex_table}},
\code{\link{example_data2}},
\code{\link{example_data3}},
\code{\link{example_data}},
\code{\link{genes_snps}},
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gtex_protein}},
\code{\link{gtex_rna}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{hpa_rna}},
\code{\link{inweb_table}},
\code{\link{irefindex_table}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c2_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_h_table}}
}
\concept{datasets}
\keyword{datasets}
