% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_markers_basic.R
\name{add_genoppi_markers}
\alias{add_genoppi_markers}
\title{Add genoppi markers}
\usage{
add_genoppi_markers(p, x = "logFC", y = "pvalue", volcano = F)
}
\arguments{
\item{p}{a ggplot.}

\item{x}{a string corresponding to column name in \code{p$data}. For plottin y axis.}

\item{volcano}{boolean. Will apply -log10 to y-axis, i.e. \code{-log10(p$data$y)}.}
}
\description{
adds plotly markers for a plot thus making the plot interactive.
}
