% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_markers_basic.R
\name{add_genoppi_markers_overlay}
\alias{add_genoppi_markers_overlay}
\title{Add genoppi markers}
\usage{
add_genoppi_markers_overlay(p, x = "logFC", y = "pvalue", volcano = F)
}
\arguments{
\item{p}{a ggplot or plotly object.}

\item{x}{a string corresponding to column name in \code{p$overlay}. For plottin y axis.}

\item{volcano}{boolean. Will apply -log10 to y-axis, i.e. \code{-log10(p$data$y)}.}
}
\description{
adds plotly markers for a plot thus making the plot interactive. Depends
on the data.frame \code{p$overlay} for the plotting.
}
