% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hover_markers.R
\name{add_layout_html_axes}
\alias{add_layout_html_axes}
\title{add reactive dashed lines}
\usage{
add_layout_html_axes(p)
}
\arguments{
\item{p}{a ggplot}
}
\description{
Format plotly axes.
}
\note{
internal
}
\seealso{
Other shiny: 
\code{\link{add_hover_lines}()},
\code{\link{add_markers_basic}()}
}
\concept{shiny}
