% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_layout_html_axes.R
\name{add_layout_html_axes_scatterplot}
\alias{add_layout_html_axes_scatterplot}
\title{layout for scatterplot}
\usage{
add_layout_html_axes_scatterplot(
  p,
  title = "",
  legend_title = bold("Overlay"),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{p}{a plotly object.}

\item{title}{the title of the plot.}

\item{legend_title}{the title of the legend.}

\item{width}{the width of the plotly object.}

\item{height}{the heigh of the plotly object.}
}
\description{
format axes of a plotly object.
}
\seealso{
Other shiny: 
\code{\link{add_hover_lines_volcano}()},
\code{\link{add_layout_html_axes_volcano}()},
\code{\link{add_markers_search_pathway}()},
\code{\link{add_markers_search}()}
}
\concept{shiny}
