% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_markers_basic_volcano.R
\name{add_layout_html_axes_volcano}
\alias{add_layout_html_axes_volcano}
\title{add reactive dashed lines}
\usage{
add_layout_html_axes_volcano(p)
}
\arguments{
\item{p}{a ggplot}
}
\description{
Format plotly axes.
}
\note{
internal
}
\seealso{
Other shiny: 
\code{\link{add_hover_lines_volcano}()},
\code{\link{add_layout_html_axes_scatterplot}()},
\code{\link{add_markers_basic_scatterplot}()},
\code{\link{add_markers_basic_volcano}()},
\code{\link{add_markers_overlay_volcano}()}
}
\concept{shiny}
