% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_line_lm.R
\name{add_line_lm}
\alias{add_line_lm}
\title{add a regression line}
\usage{
add_line_lm(
  p,
  data = NULL,
  x,
  y,
  color = "black",
  width = 1,
  dash = "dash",
  opacity = 0.9
)
}
\arguments{
\item{p}{a plotly object}

\item{data}{the data object used to draw the plotly object. By
default, the function will look for the data object within the
plotly object.}

\item{x}{for linear model, the explanatory variable.}

\item{y}{for linear model, the response variable.}

\item{color}{color of line}

\item{width}{width of line}

\item{dash}{dashed or solid?}

\item{opacity}{opacity of line.}
}
\description{
adds a regression line to plotly object
}
