% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hover_markers.R
\name{add_markers_basic}
\alias{add_markers_basic}
\title{add hover markers}
\usage{
add_markers_basic(p)
}
\arguments{
\item{plt}{a ggplot}
}
\description{
A function that takes a ggplot as input and 'plotlyfies' it so
that it becomes interactive by adding markers.
}
\note{
internal
}
\seealso{
Other shiny: 
\code{\link{add_hover_lines}()},
\code{\link{add_layout_html_axes}()}
}
\concept{shiny}
