% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_markers_basic_volcano.R
\name{add_markers_basic_volcano}
\alias{add_markers_basic_volcano}
\title{add hover markers}
\usage{
add_markers_basic_volcano(p)
}
\arguments{
\item{plt}{a ggplot}
}
\description{
A function that takes a ggplot as input and 'plotlyfies' it so
that it becomes interactive by adding markers.
}
\note{
internal
}
\seealso{
Other shiny: 
\code{\link{add_hover_lines_volcano}()},
\code{\link{add_layout_html_axes_scatterplot}()},
\code{\link{add_layout_html_axes_volcano}()},
\code{\link{add_markers_basic_scatterplot}()}
}
\concept{shiny}
