% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_markers_basic_volcano.R
\name{add_markers_overlay_volcano}
\alias{add_markers_overlay_volcano}
\title{add hover markers}
\usage{
add_markers_overlay_volcano(p, id = "SNP", id_text = "")
}
\arguments{
\item{p}{a ggplot}
}
\description{
A function that takes a ggplot as input and 'plotlyfies' it so
that it becomes interactive by adding markers to all special plots, e.g. 
a genelist or snplist.
}
\note{
internal
}
\seealso{
Other shiny: 
\code{\link{add_hover_lines_volcano}()},
\code{\link{add_layout_html_axes_scatterplot}()},
\code{\link{add_layout_html_axes_volcano}()},
\code{\link{add_markers_basic_scatterplot}()},
\code{\link{add_markers_basic_volcano}()}
}
\concept{shiny}
