% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_markers_basic.R
\name{add_markers_search}
\alias{add_markers_search}
\title{search volcano plots}
\usage{
add_markers_search(p, genes, x = "logFC", y = "pvalue", volcano = F)
}
\arguments{
\item{p}{a ggplot}

\item{genes}{a vector of genes}
}
\description{
takes a ggplot and genelist as input 
and searches the volcano plot.
}
\note{
internal
}
\seealso{
Other shiny: 
\code{\link{add_hover_lines_volcano}()},
\code{\link{add_layout_html_axes_scatterplot}()},
\code{\link{add_layout_html_axes_volcano}()}
}
\concept{shiny}
