% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_markers_search.R
\name{add_markers_search_pathway}
\alias{add_markers_search_pathway}
\title{search pathway plots}
\usage{
add_markers_search_pathway(p, pathways, mapping = NULL)
}
\arguments{
\item{p}{a ggplot}

\item{pathways}{a vector of strings containing pathway names.}

\item{mapping}{a data.frame containing a pathway column and a gene column. Will generate
a mapping between genes in the plot, and their corresponding pathways found in argument
pathways.}
}
\description{
takes a ggplot and vector as input and highligts pathways in volcano.
}
\note{
internal
}
\seealso{
Other shiny: 
\code{\link{add_hover_lines_volcano}()},
\code{\link{add_layout_html_axes_scatterplot}()},
\code{\link{add_layout_html_axes_volcano}()},
\code{\link{add_markers_search}()}
}
\concept{shiny}
