% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plotly_layout_scatter.R
\name{add_plotly_layout_scatter}
\alias{add_plotly_layout_scatter}
\title{layout for scatterplot}
\usage{
add_plotly_layout_scatter(
  p,
  title = "",
  legend_title = bold("Overlay"),
  width = NULL,
  height = NULL,
  orientation = "v",
  legend.x = 0,
  legend.y = 0
)
}
\arguments{
\item{p}{a plotly object.}

\item{title}{the title of the plot.}

\item{legend_title}{the title of the legend.}

\item{width}{the width of the plotly object.}

\item{height}{the heigh of the plotly object.}

\item{orientation}{piped to plotly. Can be 'v'ertical or 'h'orisontal.}

\item{legend.x}{shift legend by x-axis.}

\item{legend.y}{shift legend by y-axis.}
}
\description{
format axes of a plotly object.
}
\seealso{
Other plotly: 
\code{\link{add_plotly_layout_volcano}()},
\code{\link{add_plotly_line_unity}()},
\code{\link{add_plotly_threshold_lines}()},
\code{\link{add_plotly_trace}()}
}
\concept{plotly}
