% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plotly_layout_volcano.R
\name{add_plotly_layout_volcano}
\alias{add_plotly_layout_volcano}
\title{layout for volcano plot.}
\usage{
add_plotly_layout_volcano(
  p,
  height = NULL,
  width = NULL,
  legend_title = bold("Overlay"),
  orientation = "v",
  legend.x = NULL,
  legend.y = NULL
)
}
\arguments{
\item{p}{a plotly object.}

\item{height}{the heigh of the plotly object.}

\item{width}{the width of the plotly object.}

\item{legend_title}{the title of the legend.}

\item{orientation}{piped to plotly. Can be 'v'ertical or 'h'orisontal.}

\item{legend.x}{shift legend by x-axis.}

\item{legend.y}{shift legend by y-axis.}
}
\description{
Format axes.
}
\seealso{
Other plotly: 
\code{\link{add_plotly_layout_scatter}()},
\code{\link{add_plotly_line_unity}()},
\code{\link{add_plotly_threshold_lines}()},
\code{\link{add_plotly_trace}()}
}
\concept{plotly}
