% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plotly_line_unity.R
\name{add_plotly_line_unity}
\alias{add_plotly_line_unity}
\title{add unity line}
\usage{
add_plotly_line_unity(
  p,
  data = NULL,
  color = "black",
  width = 0.5,
  dash = "solid",
  opacity = 0.5
)
}
\arguments{
\item{p}{a plotly object}

\item{data}{the data object used to draw the plotly object. By
default, the function will look for the data object within the
plotly object.}

\item{color}{color of line}

\item{width}{width of line}

\item{dash}{dashed or solid?}

\item{opacity}{opacity of line.}
}
\description{
adds a unity line to plotly object
}
\seealso{
Other plotly: 
\code{\link{add_plotly_layout_scatter}()},
\code{\link{add_plotly_layout_volcano}()},
\code{\link{add_plotly_threshold_lines}()},
\code{\link{add_plotly_trace}()}
}
\concept{plotly}
