% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plotly_markers_search.R
\name{add_plotly_markers_search}
\alias{add_plotly_markers_search}
\title{search volcano plots}
\usage{
add_plotly_markers_search(p, genes, alpha = 0.5)
}
\arguments{
\item{p}{a ggplot}

\item{genes}{a vector of genes}

\item{alpha}{a numeric value between 0 (invisible) to 1 (visible).}
}
\description{
Adds search markers to a plot. Requires a plotly object.
}
\note{
internal
}
\seealso{
Other shiny: 
\code{\link{add_plotly_markers_search_pathway}()},
\code{\link{calc_logfc_limit}()},
\code{\link{get_replicate_summary_text}()},
\code{\link{parse_uploaded_file}()}
}
\concept{shiny}
