% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plotly_trace.R
\name{add_plotly_trace}
\alias{add_plotly_trace}
\title{add genoppi trace}
\usage{
add_plotly_trace(
  p,
  data,
  parameters,
  stroke_width = 0.2,
  legend = F,
  legend_group = NULL
)
}
\arguments{
\item{p}{a plot_ly() object.}

\item{data}{a data.frame with columns dataset, gene,}

\item{parameters}{a list of parameters or an environement generated with \code{environment()}.}

\item{stroke_width}{numeric. the line width/stroke of a given point.}

\item{legend}{Boolean. should legend be displayed?}

\item{legend_group}{group legend into sublegends. Argument is directly passed to \code{plotly::add_trace(legendgroup)}.}
}
\description{
adding a trace with new data to plotly is non-trivial. This function takes
care of adding data correctly with the correct coloring.
}
\note{
global variable 'global_colors' must be specified using setNames().
}
\seealso{
Other plotly: 
\code{\link{add_plotly_layout_scatter}()},
\code{\link{add_plotly_layout_volcano}()},
\code{\link{add_plotly_line_unity}()},
\code{\link{add_plotly_threshold_lines}()}
}
\concept{plotly}
