% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{as.bait}
\alias{as.bait}
\title{as.bait}
\usage{
as.bait(bait)
}
\arguments{
\item{bait}{string indicating the bait.}
}
\description{
quickly format a gene as a bait so that it can be used by various overlay functions.
This functional will create a named data.frame with column 'gene', 'col_significant' and
'col_other', that can be directly inputted into plot_overlay.
}
\examples{
\dontrun{
# overlay the bait
example_data \%>\%
  calc_mod_ttest() \%>\%
  id_significant_proteins() \%>\%
  plot_volcano_basic() \%>\%
  plot_overlay(as.bait('BCL2')) \%>\%
}
}
\seealso{
Other misc: 
\code{\link{\%nin\%}()},
\code{\link{as.goi}()},
\code{\link{bold}()},
\code{\link{catf}()},
\code{\link{color_distinct}()},
\code{\link{color_gradient}()},
\code{\link{enumerate_replicate_combinations}()},
\code{\link{get_gg_legend}()},
\code{\link{hline}()},
\code{\link{italics}()},
\code{\link{lun}()},
\code{\link{null_omit}()},
\code{\link{shape_to_symbol}()},
\code{\link{symbol_to_shape}()},
\code{\link{vline}()}
}
\concept{misc}
