% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathways.R
\name{as.geneset}
\alias{as.geneset}
\title{plot pathways}
\usage{
as.geneset(data, database)
}
\arguments{
\item{database}{see get_pathways}

\item{pulldown}{pulldown data}
}
\description{
similar to \code{?plot_overlay}.
}
\note{
this is a work in progress.
}
\keyword{internal}
