% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{as.goi}
\alias{as.goi}
\title{as gene of interest}
\usage{
as.goi(
  genes,
  col_significant = "cyan",
  col_other = "grey",
  shape = 21,
  dataset = "GOI"
)
}
\arguments{
\item{genes}{string indicating the bait.}

\item{col_significant}{color of significant interactor.}

\item{col_other}{color of non-significant interactor.}

\item{shape}{numeric. 21 is default for circle.}

\item{dataset}{used internally in genoppi to plot multiple datasets.}
}
\description{
quickly format the gene so that it can be used by various overlay functions.
This functional will create a named data.frame with column 'gene', 'col_significant' and
'col_other', that can be directly inputted into plot_overlay.
}
\examples{
\dontrun{
# overlay the bait
example_data \%>\%
 calc_mod_ttest() \%>\%
 id_significant_proteins() \%>\%
 plot_volcano_basic() \%>\%
 plot_overlay(as.goi(c('BCL2', 'FUS', 'TRIM28')))
}
}
\seealso{
Other misc: 
\code{\link{\%nin\%}()},
\code{\link{as.bait}()},
\code{\link{bold}()},
\code{\link{catf}()},
\code{\link{color_distinct}()},
\code{\link{color_gradient}()},
\code{\link{enumerate_replicate_combinations}()},
\code{\link{get_gg_legend}()},
\code{\link{hline}()},
\code{\link{italics}()},
\code{\link{lun}()},
\code{\link{null_omit}()},
\code{\link{shape_to_symbol}()},
\code{\link{symbol_to_shape}()},
\code{\link{vline}()}
}
\concept{misc}
