% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathways.R
\name{as.pathway}
\alias{as.pathway}
\title{plot pathways}
\usage{
as.pathway(data, database)
}
\arguments{
\item{database}{see get_pathways}

\item{pulldown}{pulldown data}
}
\description{
similar to \code{?plot_overlay}.
}
\note{
this is a wark in progress.
}
\keyword{internal}
