% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{assign_color}
\alias{assign_color}
\title{assign color by column}
\usage{
assign_color(df, col, order_by_freq = T, palette = color_distinct())
}
\arguments{
\item{df}{a data.frame}

\item{col}{the identifiying column for assiging color}

\item{order_by_freq}{boolean. Should the data be order by frequency of col?}

\item{palette}{color palette, i.e. a vector of colors.}
}
\description{
assigns color to a data.frame by a certain column
}
