% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{assign_freq}
\alias{assign_freq}
\title{assign frequency}
\usage{
assign_freq(df, col)
}
\arguments{
\item{df}{a data.frame}

\item{col}{the column which to assign frequency to.}
}
\description{
assign a frequency of occurences to a dataframe
}
