% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_volcano.R
\name{basic_volcano}
\alias{basic_volcano}
\title{Plot basic volcano}
\usage{
takes in a data.frane with the columns gene, logFC, pvalue and significant 
to draw a volcano. Optionally, a column indicating 'color' (string) and 'label' (boolean),
can be supplied to indicate the volcano color scheme and whether to draw names of proteins.
}
\arguments{
\item{df}{a data.frame with at least columns gene, logFC, pvalue and signifidant.}

\item{bait}{the gene name of the bait.}

\item{title}{the title of the plot \code{?ggplot2::labs}.}
}
\description{
A function that draw a basic volcano plot.
}
