% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bioplex_table}
\alias{bioplex_table}
\title{BioPlex protein-protein interactions}
\format{
A tibble object with 159,422 rows and 4 columns:
\describe{
  \item{Gene1}{<chr> Gene name of first protein}
  \item{Gene2}{<chr> Gene name of second protein}
  \item{Score}{<chr> Interaction confidence score (i.e. pInt in BioPlex data),
separated by "," if interaction found in both cell lines} 
  \item{CellLine}{<chr> Cell line(s) in which the interaction was identified}
}
}
\source{
BioPlex 3.0 data downloaded on 2023-12-05 from:
\url{https://bioplex.hms.harvard.edu/interactions.php}
}
\usage{
bioplex_table
}
\description{
A dataset containing 159,422 undirected protein-protein interactions for
14,145 human genes in BioPlex 3.0, identified by AP-MS in HEK293T and/or HCT116 cells.
}
\references{
Huttlin EL, Bruckner RJ, Navarrete-Perea J, et al. Dual proteome-scale networks
reveal cell-specific remodeling of the human interactome. 
Cell. 2021;184(11):3022-3040.e28. doi:10.1016/j.cell.2021.04.011
}
\seealso{
Other datasets: 
\code{\link{accession_gene_table}},
\code{\link{example_data2}},
\code{\link{example_data3}},
\code{\link{example_data}},
\code{\link{genes_snps_hash}},
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gtex_brain_table}},
\code{\link{gtex_protein_table}},
\code{\link{gtex_rna_table}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{hpa_rna_table}},
\code{\link{inweb_table}},
\code{\link{irefindex_table}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c2_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_c8_table}},
\code{\link{msigdb_h_table}},
\code{\link{string_table}}
}
\concept{datasets}
\keyword{datasets}
