% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_adujusted_enrichment.R
\name{calc_adjusted_enrichment}
\alias{calc_adjusted_enrichment}
\title{calculate adjusted enrichment}
\usage{
calc_adjusted_enrichment(data, reference, col.by = "tissue", bait = NULL)
}
\arguments{
\item{data}{proteomic data with gene and significant columns.}

\item{reference}{table with col.by, gene, significant in columns.}

\item{col.by}{string. What column contains the group?}

\item{bait}{bait. See \code{?calc_hyper}.}
}
\description{
calculate benjamin-hochberg adjusted enrichment of genes in tissue.
}
