% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fisher.R
\name{calc_fisher}
\alias{calc_fisher}
\title{Perform Fisher's exact test}
\usage{
calc_fisher(df, listDf, intersectDf, bait = NULL)
}
\arguments{
\item{df}{data.frame containing proteomic data, with gene and significant columns}

\item{listDf}{data.frame containing data to overlap, with listName, gene and significant columns}

\item{intersectDf}{data.frame contaning listName and intersectN columnsl intersectN = vector of boolean variables indicating if total population should be intersect of the two datasets}

\item{bait}{name of bait protein}
}
\value{
list of data.frame and list. Data.frame contains list_name, overlap_count, dfOnly_count, listOnly_count, neither_count, population_count, pvalue. List of lists (for each listName) contains genes names corresponding to each group (overlap_genes, dfOnly_genes, listOnly_genes, neither_genes, population_genes).
}
\description{
Use one-tailed Fisher's exact test to calcualte overlap enrichment between proteomic data and other datasets (e.g. InWeb, gene lists, etc.)
}
