% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hyper.R
\name{calc_hyper}
\alias{calc_hyper}
\title{Perform hypergeometric test}
\usage{
calc_hyper(df, listName, listDf, intersectN, bait = NULL)
}
\arguments{
\item{df}{data.frame containing proteomic data, with gene and significant columns}

\item{listName}{name of data to overlap}

\item{listDf}{data.frame containing data to overlap, with gene and significant columns}

\item{intersectN}{boolean variable indicating if total population should be intersect of the two datasets}

\item{bait}{name of bait protein}
}
\value{
list of data.frame and list. Data.frame contains list_name, successInSample_count (x), sample_count (n), notSample_count (N-n), success_count (k), pvalue. List contains genes names corresponding to each group (successInSample_genes, sample_genes, notSample_genes, success_genes).
}
\description{
Use one-tailed hypergeometric test to calcualte overlap enrichment between proteomic data and another dataset (e.g. InWeb, gene list, etc.)
}
