% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_logfc.R
\name{calculate_logfc}
\alias{calculate_logfc}
\title{calculate log fold change (LogFC)}
\usage{
calculate_logfc(df)
}
\arguments{
\item{df}{a data.frame with bait and controls}
}
\description{
calculates log fold change by substracting a case column with a bait column. Will apply
to all numeric entries with the rep1-9 in the column names
}
\note{
assummes log2(bait1), log2(control1), log2(bait2)..
}
\author{
flassen
}
