% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{catf}
\alias{catf}
\title{warnings to stderr}
\usage{
catf(msg, file = stderr())
}
\arguments{
\item{msg}{the message}

\item{file}{string, e.g. a filename.}
}
\description{
sends a message to stderr (i.e shiny)
}
