% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_labels.R
\name{collapse_labels}
\alias{collapse_labels}
\title{collaps overlay data}
\usage{
collapse_labels(
  overlay,
  dataset = "dataset",
  collapse_into = "alt_label",
  collapse_by = "gene",
  dataset_collapse_sep = ": ",
  item_sep = " <br>"
)
}
\arguments{
\item{overlay}{a data.frame}

\item{dataset}{the main dataset name that is to be combined.}

\item{collapse_into}{the column that will contain the collapsed entries.}

\item{collapse_by}{what identifying colummn is the data to be collapsed by?}

\item{dataset_collapse_sep}{how should the dataset column and collapse-column 
be seperated by when collapsed?}

\item{item_sep}{how should each collapsed item be seperated from the next? Default is
newline.}
}
\description{
a data.frame used to overlay items in genoppi
will look for non-unqiue gene names and combine them into 
a single informative line using their alt_label (e.g. snp)
and their corrsponding dataset.
}
