% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{color_distinct}
\alias{color_distinct}
\title{distinct coloring}
\usage{
color_distinct(length.out = 74)
}
\arguments{
\item{length.out}{repeats the vector.}
}
\description{
generates vector of 74 distinct colors from RColorBrewer.
}
\seealso{
Other misc: 
\code{\link{\%nin\%}()},
\code{\link{as.bait}()},
\code{\link{as.goi}()},
\code{\link{bold}()},
\code{\link{catf}()},
\code{\link{color_gradient}()},
\code{\link{enumerate_replicate_combinations}()},
\code{\link{get_gg_legend}()},
\code{\link{hline}()},
\code{\link{italics}()},
\code{\link{lun}()},
\code{\link{null_omit}()},
\code{\link{shape_to_symbol}()},
\code{\link{symbol_to_shape}()},
\code{\link{table_symbols}()},
\code{\link{vline}()}
}
\concept{misc}
