% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_interactive.R
\name{combine_dataset_and_significance}
\alias{combine_dataset_and_significance}
\title{Combine data set and significance}
\usage{
combine_dataset_and_significance(data, sig_text = "", insig_text = "(N)")
}
\arguments{
\item{data}{a data or overlay data.frame with \code{dataset} as column.}

\item{sig_text}{what should be the significance text?}

\item{insig_text}{what should be the insignificance text?}
}
\description{
In order to correctly draw plotly points
the identifier (dataset) must have a unique colors associated with it.
This function appends the dataset column with a significance text.
}
