% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_genoppi_venn.R
\name{draw_genoppi_venn}
\alias{draw_genoppi_venn}
\title{Draw venn diagrams}
\usage{
draw_genoppi_venn(
  x,
  main = "",
  colors = genoppi::color_distinct(),
  alpha = NULL,
  cat.cex = 1.1,
  cat.dist = 0.05,
  sub.pos = 0,
  fill = NULL
)
}
\arguments{
\item{x}{a list of two vectors that contain some overlapping strings.}

\item{main}{title, typically a p-value.}

\item{colors}{color scheme.}

\item{alpha}{numeric, transparency value between 0 and 1.}

\item{cat.cex}{numeric. the size of category labels}

\item{cat.dist}{the border distance between venn diagrams and labels}

\item{sub.pos}{numeric. the position of the labels.}

\item{fill}{the fill color of the venn diagrams.}
}
\description{
drwas a simple 1 by 1 overlap venn diagram.
}
\note{
NAs in x will be discarded and produce a warning.
}
