% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate_replicate_combinations.R
\name{enumerate_replicate_combinations}
\alias{enumerate_replicate_combinations}
\title{enumerate replicate combinations}
\usage{
enumerate_replicate_combinations(df)
}
\arguments{
\item{df}{the data frame from which replicate combinations are extracted}
}
\description{
enumerates unique replicate comparisons, extract rep-rep,
  sample-sample, control-control, and sample-control combinations
}
\seealso{
Other misc: 
\code{\link{\%nin\%}()},
\code{\link{as.bait}()},
\code{\link{as.goi}()},
\code{\link{bold}()},
\code{\link{catf}()},
\code{\link{color_distinct}()},
\code{\link{color_gradient}()},
\code{\link{get_gg_legend}()},
\code{\link{hline}()},
\code{\link{italics}()},
\code{\link{lun}()},
\code{\link{null_omit}()},
\code{\link{shape_to_symbol}()},
\code{\link{symbol_to_shape}()},
\code{\link{vline}()}
}
\concept{misc}
