% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{find_docs}
\alias{find_docs}
\title{find docs}
\usage{
find_docs(file = "inweb_table.info", dir = "documentation")
}
\arguments{
\item{file}{file / regex pattern}

\item{dir}{directory}
}
\description{
find documentation objects for shiny app.
}
