% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{findit}
\alias{findit}
\title{findit}
\usage{
findit(what = "name", directory = "R")
}
\arguments{
\item{what}{string to find}

\item{directory}{string, directory}
}
\description{
find files and lines that contain 'what'
}
\note{
for debugging r-package.
}
\keyword{internal}
