% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{findnamespace}
\alias{findnamespace}
\title{find namespace}
\usage{
findnamespace(ns = ls(getNamespace("ggplot2")), directory = "R")
}
\arguments{
\item{directory}{string, directory}
}
\description{
find all scrips containing namespace objects
}
\note{
for debugging r-package.
}
\keyword{internal}
