% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes_snps}
\alias{genes_snps}
\title{HuRI data (HI-union network)}
\format{
TO ADD
\describe{
  \item{Gene1}{HGNC gene symbol of first protein}
  \item{Gene2}{HGNC gene symbol of second protein}
}

A hash object with each of 18,116 genes as key and a vector of SNPs as value
}
\source{
HuRI data (HI-union network) data downloaded on 2023-10-10 from:
\url{https://www.nature.com/articles/s41586-020-2188-x}
TODO: try direct download from https://static-content.springer.com/esm/art%3A10.1038%2Fs41586-020-2188-x/MediaObjects/41586_2020_2188_MOESM3_ESM.zip

snp_to_gene.Rdata file created by April Kim.
}
\usage{
genes_snps
}
\description{
TO ADD

A data set containing 1000 Genomes phase 3 single-nucleotide polymorphsims (SNPs) mapped to 
18,116 gene symbols, using linkage disequilibrium information from the CEU population. See cited Genoppi
reference for details on mapping method.
}
\details{
This dataset was generated from the source BioPlex database by excluding all non-binary and non-human
interactions. HGNC gene symbols were then extracted from the database along with the relevant score columns.
Rows corresponding to the same gene pairs were merged into a single row by keeping the maximum scores.
}
\references{
Luck, K., Kim, DK., Lambourne, L. et al. A reference map of the human binary protein interactome. Nature 580, 402–408
(2020). https://doi.org/10.1038/s41586-020-2188-x

1000 Genomes Project Consortium, Auton A, Brooks LD, et al. A global reference for human genetic variation.
Nature. 2015;526(7571):68-74. doi:10.1038/nature15393  

Pintacuda G, Lassen FH, Hsu YH, Kim A, Martín JM, Malolepsza E et al. Genoppi is an open-source software
for robust and standardized integration of proteomic and genetic data. Nat Commun. 2021;12(1):2580.
doi:10.1038/s41467-021-22648-5
}
\seealso{
Other datasets: 
\code{\link{accession_gene_table}},
\code{\link{bioplex_table}},
\code{\link{example_data2}},
\code{\link{example_data3}},
\code{\link{example_data}},
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gtex_brain_table}},
\code{\link{gtex_protein_table}},
\code{\link{gtex_rna_table}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{hpa_rna_table}},
\code{\link{inweb_table}},
\code{\link{irefindex_table}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c2_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_c8_table}},
\code{\link{msigdb_h_table}}

Other datasets: 
\code{\link{accession_gene_table}},
\code{\link{bioplex_table}},
\code{\link{example_data2}},
\code{\link{example_data3}},
\code{\link{example_data}},
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gtex_brain_table}},
\code{\link{gtex_protein_table}},
\code{\link{gtex_rna_table}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{hpa_rna_table}},
\code{\link{inweb_table}},
\code{\link{irefindex_table}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c2_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_c8_table}},
\code{\link{msigdb_h_table}}
}
\concept{datasets}
\keyword{datasets}
