% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bioplex_list.R
\name{get_bioplex_list}
\alias{get_bioplex_list}
\title{Retrieve Bioplex 3.0 interactors for a given bait}
\usage{
get_bioplex_list(bait, score = 0)
}
\arguments{
\item{bait}{string. name of bait protein}

\item{score}{numeric. confidence score (between 0 and 1) cutoff. See \code{?bioplex_table}.}
}
\value{
data.frame containing gene and significant columns for all non-bait BioPlex genes 
(significant=T for BioPlex 3.0 interactors of bait). NULL if bait not found in BioPlex.
}
\description{
Use bioplex_table data to get BioPlex interactors and non-interactors of bait. 
See \code{?bioplex_table} for more details about the data set.
}
\examples{
\dontrun{
df1 <- get_bioplex_list('BCL2',score = 0.9)
}
}
