% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_from_snp.R
\name{get_gene_from_snp}
\alias{get_gene_from_snp}
\title{get the genes that matches a vector of SNPs}
\usage{
get_gene_from_snp(snp, invert = F)
}
\arguments{
\item{snp}{a vector or character contain snp names (rsIDs)}

\item{invert}{boolean. Inverts the output, so that a list of SNPs is outputted instead.}
}
\value{
a list of genes (if invert is FALSE)
}
\description{
Create gene list data.frame from input SNP lists
}
\details{
Applies the SNP-to-gene mapping described in the paper. Please, see
the paper for details about reference panel and ancestry.
}
\examples{
\dontrun{
get_gene_from_snp('rs12516') # BRCA1
}
}
