% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_list.R
\name{get_gene_list}
\alias{get_gene_list}
\title{Read in gene list data}
\usage{
get_gene_list(filename, header = T, sep = "\\t")
}
\arguments{
\item{filename}{file path of gene list file (containing gene and (optional) significant columns)}

\item{header}{boolean indicating whether gene list file contains header}

\item{sep}{file delimiter}
}
\value{
list of (data.frame containing gene and significant columns, boolean value indicating if non-significant genes are included (for calling overlap enrichement functions))
}
\description{
Create gene list data.frame from input file
}
