% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_lists.R
\name{get_gene_lists}
\alias{get_gene_lists}
\title{Read in gene lists}
\usage{
get_gene_lists(filename, sep = "\\t")
}
\arguments{
\item{filename}{file path of gene list file (contain headers: listName, gene and (optional) significant columns)}

\item{sep}{file delimiter}
}
\value{
list of two data.frames. First data.frame contains listName, gene and significant columns. Second data.frame contains listName and intersectN columns; intersectN = boolean variable indicating if non-significant genes are included in each list (for calling overlap enrichement functions))
}
\description{
Create gene list data.frame from input file
}
