% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_gg_legend}
\alias{get_gg_legend}
\title{get ggplot legend}
\usage{
get_gg_legend(a.gplot)
}
\arguments{
\item{a.gplot}{a ggplot}
}
\description{
gets a ggplot legend
}
\note{
code snippet from \url{https://stackoverflow.com/questions/12041042/how-to-plot-just-the-legends-in-ggplot2}
}
\seealso{
Other misc: 
\code{\link{\%nin\%}()},
\code{\link{as.bait}()},
\code{\link{bold}()},
\code{\link{catf}()},
\code{\link{color_distinct}()},
\code{\link{color_gradient}()},
\code{\link{enumerate_replicate_combinations}()},
\code{\link{italics}()},
\code{\link{lun}()},
\code{\link{null_omit}()},
\code{\link{shape_to_symbol}()},
\code{\link{symbol_to_shape}()}
}
\concept{misc}
