% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_interactive.R
\name{get_global_colors}
\alias{get_global_colors}
\title{get global colors}
\usage{
get_global_colors(data, overlay)
}
\arguments{
\item{data}{a data.frame object.}

\item{overlay}{a data.frame object.}
}
\description{
the colors supplied to plotly must be in a global colors
scheme with appropiate names (similar to a dict). This function generated
this mapping, so that the color scheme can be correctly plotted.
}
