% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gnomad_constraints.R
\name{get_gnomad_constraints}
\alias{get_gnomad_constraints}
\title{get gnomad constraint table}
\usage{
get_gnomad_constraints(gene)
}
\arguments{
\item{gene}{a string}
}
\description{
returns the gnomad constraint table based on gnomad 2.1.1
}
