% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gwas_list.R
\name{get_gwas_list}
\alias{get_gwas_list}
\title{Retreive GWAS catalog data for a given list of traits}
\usage{
get_gwas_list(traits, genes)
}
\arguments{
\item{traits}{vector of GWAS catalog traits}

\item{genes}{vector of gene names (genes detected in proteomic data)}
}
\value{
list of two data.frames. First data.frame contains gene and SNP columns (restricting to genes in input genes vector; genes and SNPs can appear multiple times). Second data.frame contains GWAS catalog info for the included SNPs.
}
\description{
Create gene list data.frame from input traits
}
