% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gwas_lists.R
\name{get_gwas_lists}
\alias{get_gwas_lists}
\title{Retreive GWAS catalog data for a given list of traits}
\usage{
get_gwas_lists(traits, genes)
}
\arguments{
\item{traits}{vector of GWAS catalog traits. See \code{gwas_table$DISEASE.TRAIT}}

\item{genes}{vector of gene names (genes detected in proteomic data)}
}
\value{
list of two data.frames. First data.frame contains trait, gene and SNP columns (restricting to genes in input genes vector). Second data.frame contains GWAS catalog info for the included SNPs.
}
\description{
Create gene list data.frame from input traits
}
