% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hpa_lists.R
\name{get_hpa_lists}
\alias{get_hpa_lists}
\title{Retreive Human Protein Atlas specific data}
\usage{
get_hpa_lists(tissue = NULL, genes = NULL)
}
\arguments{
\item{tissue}{vector of tissue names (character). See details for acceptable names.}

\item{genes}{vector of gene names (genes detected in proteomic data)}
}
\description{
Create gene list data.frame from input traits
}
\details{
The following tissues are available:
\itemize{
\item liver
\item intestine
\item lung
\item esophagus
\item lymphoid tissue
\item tongue
\item gallbladder
\item pancreas
\item stomach 1
\item adrenal gland
\item placenta
\item skin 1
\item adipose tissue
\item epididymis
\item vagina
\item parathyroid gland
\item skeletal muscle
\item retina
\item testis
\item smooth muscle
\item brain
\item kidney
\item ovary
\item bone marrow
\item breast
\item prostate
\item pituitary gland
\item ductus deferens
\item seminal vesicle
\item blood
\item heart muscle
\item cervix, uterine
\item endometrium 1
\item salivary gland
\item fallopian tube
\item thyroid gland
\item urinary bladder
}
}
