% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_huri_list.R
\name{get_huri_list}
\alias{get_huri_list}
\title{Retrieve HuRI interactors for a given bait}
\usage{
get_huri_list(bait)
}
\arguments{
\item{bait}{string. name of bait protein}
}
\value{
data.frame containing gene and significant columns for all non-bait bioplex genes 
(significant=T for huri interactors of bait). NULL if bait not found in huri.
}
\description{
Use huri_table data to get huri interactors and non-interactors of bait.
See \code{?huri_table} for more details about the data set.
}
\examples{
\dontrun{
df1 <- get_huri_list('BCL2')
}
}
