% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_irefindex_list.R
\name{get_irefindex_list}
\alias{get_irefindex_list}
\title{Retrieve irefindex for a given bait}
\usage{
get_irefindex_list(bait, score = 1)
}
\arguments{
\item{bait}{string. name of bait protein}

\item{score}{numeric. confidence score (number of publications) cutoff.}
}
\value{
data.frame containing gene and significant columns for all non-bait iRefIndex genes 
(significant=T for iRefIndex interactors of bait). NULL if bait not found in iRefIndex.
}
\description{
Use irefindex_table data to get iRefIndex interactors and non-interactors of bait. 
See \code{?irefindex_table} for more details about the data set.
}
\examples{
\dontrun{
df1 <- get_irefindex_list('BCL2',score = 2)
}
}
