% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_overlay_df.R
\name{get_overlay_df}
\alias{get_overlay_df}
\title{get overlay data.frame}
\usage{
get_overlay_df(
  p,
  reference,
  match = "gene",
  sig_text = NULL,
  insig_text = NULL,
  legend_nchar_max = NULL,
  nchar_max_collapse = NULL,
  dont_inherit = c("dataset", "color", "size", "size_gg", "shape", "group",
    "col_border")
)
}
\arguments{
\item{p}{A ggplot object. Usually passed down from \code{plot_volcano_basic} or \code{plot_scatter_basic}.}

\item{reference}{a list of data.frames that are preferably named. The name of the list will passed down to
the data.frame as the column 'dataset'. Alternatively, the dataset can have a column name dataset.}

\item{match}{by what string should the ggplot and overlay be merged? Default is 'gene'.}

\item{sig_text}{string. text for enriched interactors to be displayed in legend.}

\item{insig_text}{string. Text for non-enriched interactors to be displayed in legend.}

\item{legend_nchar_max}{maximum amount of allowed characters in the legend.}

\item{nchar_max_collapse}{what charcter should be used for line break? Default is HTML line break \code{"<br>".}}

\item{dont_inherit}{the parameters/columns that should not be inherited from the plotting environment.}
}
\description{
Uses a ggplot and a reference data.frame to generate
'match' genes that are present in both the plot and the reference data.
 Subsequently, plot information for the individual points are extracted
 and can now be used as an overlay. See {?plot_overlay}
}
