% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pcnet_list.R
\name{get_pcnet_list}
\alias{get_pcnet_list}
\title{Retrieve Bioplex 3.0 interactors for a given bait}
\usage{
get_pcnet_list(bait)
}
\arguments{
\item{bait}{string. name of bait protein}
}
\value{
data.frame containing gene and significant columns for all non-bait PCNet genes
(significant=T for PCNet interactors of bait). NULL if bait not found in PCNet.
}
\description{
Use pcnet_table data to get PCNet interactors and non-interactors of bait.
See \code{?pcnet_table} for more details about the data set.
}
\examples{
\dontrun{
df1 <- get_pcnet_list('BCL2')
}
}
