% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ppi_list.R
\name{get_ppi_list}
\alias{get_ppi_list}
\title{Retrieve PPI data for a given bait}
\usage{
get_ppi_list(bait, dataset)
}
\arguments{
\item{bait}{name of bait protein}

\item{dataset}{one of "bioplex", "inweb", "irefindex", "string"}
}
\value{
data.frame containing gene and significant columns for all non-bait genes in
specific dataset (significant=T for interactors of bait). NULL if bait not found in dataset.
}
\description{
Get interactors and non-interactors of bait from specified PPI dataset.
}
\details{
this function does not use confidence score cutoffs to filter interactions.
Use \code{?get_bioplex_list}, \code{?get_inweb_list}, \code{?get_irefindex_list}, and 
\code{?get_string_list} instead to set cutoffs.
}
\examples{
\dontrun{
df1 <- get_ppi_list('BCL2','inweb')
}
}
