% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_replicate_summary_text.R
\name{get_replicate_summary_text}
\alias{get_replicate_summary_text}
\title{get replicate summary}
\usage{
get_replicate_summary_text(thresholds, data, reps, digits = 3)
}
\arguments{
\item{thresholds}{a list in the format of thresholds$sig$sig and thresholds$fc$sig.}

\item{data}{proteomic data. Must contain a significant column.}

\item{reps}{a names list of replicate correlations.}

\item{digits}{how many digits should the table be rounded with?}
}
\description{
a function used by shiny to generate summary tables over replicates.
}
\seealso{
Other shiny: 
\code{\link{add_plotly_markers_search_pathway}()},
\code{\link{add_plotly_markers_search}()},
\code{\link{calc_logfc_limit}()},
\code{\link{parse_uploaded_file}()}
}
\concept{shiny}
