% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shiny_errors.R
\name{get_shiny_errors}
\alias{get_shiny_errors}
\title{get shiny errors}
\usage{
get_shiny_errors(df)
}
\arguments{
\item{df}{a data.frame}
}
\description{
parses a data.frame and returns a string potential problems that might occour downstream.
}
