% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snp_lists.R
\name{get_snp_lists}
\alias{get_snp_lists}
\title{Map SNP lists to genes}
\usage{
get_snp_lists(infile, genes)
}
\arguments{
\item{infile}{a data.frame or a file path of SNP lists (two columns with header: listName, SNP)}

\item{genes}{vector of gene names (genes detected in proteomic data)}
}
\value{
data.frame containing listName, gene and SNP columns (restricting to genes in input genes vector).
}
\description{
Create gene list data.frame from input SNP lists
}
