% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_string_list.R
\name{get_string_list}
\alias{get_string_list}
\title{Retrieve STRING interactors for a given bait}
\usage{
get_string_list(bait)
}
\arguments{
\item{bait}{string. name of bait protein}
}
\value{
data.frame containing gene and significant columns for all non-bait STRING genes
(significant=T for STRING interactors of bait). NULL if bait not found in STRING.
}
\description{
Use string_table data to get STRING interactors and non-interactors of bait.
See \code{?string_table} for more details about the data set.
}
\examples{
\dontrun{
df1 <- get_string_list('BCL2')
}
}
