% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tissue_list.R
\name{get_tissue_lists}
\alias{get_tissue_lists}
\title{deprecated. See ?get_tissue_list}
\usage{
get_tissue_lists(tissue, table = gtex_rna_table)
}
\arguments{
\item{tissue}{deprecated. See ?get_tissue_list}

\item{table}{deprecated. See ?get_tissue_list}
}
\description{
deprecated. See ?get_tissue_list
}
