% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gnomad_table}
\alias{gnomad_table}
\title{gnomAD gene contstraint data}
\format{
A tibble object with 16,857 rows and 5 columns:
\describe{
  \item{gene}{<chr> gene name}
  \item{pLI}{<dbl> probability of loss-of-function intolerance computed from pLoF variants (higher values = more constrained)} 
  \item{LOEUF}{<dbl> upper bound of 90-percent confidence interval for observed/expected ratio for pLoF variants (lower values = more constrained)}
  \item{MisZ}{<dbl> Z-score for missense variants (higher values = more constrained)}
  \item{SynZ}{<dbl> Z-score for synonymous variants (higher values = more constrained)}
}
}
\source{
gnomAD v4 gene constraint data downloaded from:
\url{https://gnomad.broadinstitute.org/downloads#v4-constraint}
}
\usage{
gnomad_table
}
\description{
gnomAD gene constraint scores for 16,857 protein-coding genes.
}
\references{
Karczewski KJ, Francioli LC, Tiao G, et al. The mutational constraint spectrum quantified from variation in 141,456 humans.
Nature. 2020;581(7809):434-443. doi:10.1038/s41586-020-2308-7 

Chen S, Francioli LC, Goodrich JK, et al. A genome-wide mutational constraint map quantified from variation in 76,156 human genomes.
bioRxiv. 2022. doi:10.1101/2022.03.20.485034

See contstraint scores documentation at: https://gnomad.broadinstitute.org/help/constraint
}
\seealso{
Other datasets: 
\code{\link{accession_gene_table}},
\code{\link{bioplex_table}},
\code{\link{example_data2}},
\code{\link{example_data3}},
\code{\link{example_data}},
\code{\link{genes_snps_hash}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gtex_brain_table}},
\code{\link{gtex_protein_table}},
\code{\link{gtex_rna_table}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{hpa_rna_table}},
\code{\link{huri_table}},
\code{\link{inweb_table}},
\code{\link{irefindex_table}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c2_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_c8_table}},
\code{\link{msigdb_h_table}},
\code{\link{pcnet_table}},
\code{\link{string_table}}
}
\concept{datasets}
\keyword{datasets}
