% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goa_cc_table}
\alias{goa_cc_table}
\title{GO cellular component annotations}
\format{
A tibble object with 179,867 rows and 3 columns:
\describe{
  \item{Gene.symbol}{<chr> HGNC gene symbol}
  \item{GO.ID}{<chr> GO term ID}
  \item{GO.name}{<chr> GO term name}
}
}
\source{
Homo sapiens EBI Gene Ontology Annotation Database protein dataset (2023-07-27 release) downloaded from:
\url{http://current.geneontology.org/products/pages/downloads.html}
}
\usage{
goa_cc_table
}
\description{
A dataset containing annotations for 18,868 genes in 1,800 GO cellular component terms.
}
\references{
Ashburner M, Ball CA, Blake JA, et al. Gene ontology: tool for the unification of biology.
The Gene Ontology Consortium. Nat Genet. 2000;25(1):25-29. doi:10.1038/75556  

Gene Ontology Consortium, Aleksander SA, Balhoff J, et al. The Gene Ontology knowledgebase in 2023.
Genetics. 2023;224(1):iyad031. doi:10.1093/genetics/iyad031
}
\seealso{
Other datasets: 
\code{\link{accession_gene_table}},
\code{\link{bioplex_table}},
\code{\link{example_data2}},
\code{\link{example_data3}},
\code{\link{example_data}},
\code{\link{genes_snps}},
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_mf_table}},
\code{\link{gtex_protein}},
\code{\link{gtex_rna}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{hpa_rna}},
\code{\link{inweb_table}},
\code{\link{irefindex_table}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c2_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_h_table}}
}
\concept{datasets}
\keyword{datasets}
