% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goa_mf_table_data.R
\docType{data}
\name{goa_mf_table}
\alias{goa_mf_table}
\title{GOA molecular function annotation table}
\format{data.frame object with: Gene.symbol, GO.ID, GO.name}
\source{
"GOA Homo sapiens protein dataset, 2020-03-23 release. Downloaded on 2020-03-30 (http://current.geneontology.org/products/pages/downloads.html)."
}
\usage{
goa_mf_table
}
\description{
A table containing 65,891 GO molecular function annotations (17,654 unique gene symbols, 4,271 unique GO MF terms).
}
\seealso{
Other datasets: 
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{msigdb_table}}
}
\concept{datasets}
\keyword{datasets}
