% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{\%nin\%}
\alias{\%nin\%}
\title{not in}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{value x}

\item{y}{value y}
}
\description{
returns true for x not in y
}
\seealso{
Other misc: 
\code{\link{as.bait}()},
\code{\link{bold}()},
\code{\link{catf}()},
\code{\link{color_distinct}()},
\code{\link{color_gradient}()},
\code{\link{enumerate_replicate_combinations}()},
\code{\link{italics}()},
\code{\link{lun}()},
\code{\link{null_omit}()}
}
\concept{misc}
