% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtex_brain_table}
\alias{gtex_brain_table}
\title{GTEx RNA tissue specificity annotations for brain regions/tissues}
\format{
A tibble object with 228,657 rows and 3 columns:
\describe{
  \item{tissue}{<chr> Tissue name}
  \item{gene}{<chr> HGNC gene symbol}
  \item{significant}{<lgl> whether or not the gene is tissue-specific} 
}
}
\source{
GTEx_brain.tstat.tsv downloaded on 2023-10-13 from:
\url{https://console.cloud.google.com/storage/browser/broad-alkesgroup-public-requester-pays/LDSCORE/LDSC_SEG_ldscores/tstats/}
}
\usage{
gtex_brain_table
}
\description{
A dataset containing annotations for 17,589 tissue-specific genes in 13 brain tissues derived from
GTEx RNA-seq data. See references for definition of tissue-specific genes. In brief, genes that rank
in the top decile of a t-statistic distribution are defined as the tissue-specific genes in each tissue.
}
\references{
Finucane HK, Reshef YA, Anttila V, et al. Heritability enrichment of specifically expressed genes identifies
disease-relevant tissues and cell types. Nat Genet. 2018;50(4):621-629. doi:10.1038/s41588-018-0081-4
}
\seealso{
Other datasets: 
\code{\link{accession_gene_table}},
\code{\link{bioplex_table}},
\code{\link{example_data2}},
\code{\link{example_data3}},
\code{\link{example_data}},
\code{\link{genes_snps_hash}},
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gtex_protein_table}},
\code{\link{gtex_rna_table}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{hpa_rna_table}},
\code{\link{inweb_table}},
\code{\link{irefindex_table}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c2_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_c8_table}},
\code{\link{msigdb_h_table}},
\code{\link{string_table}}
}
\concept{datasets}
\keyword{datasets}
