% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtex_protein_table}
\alias{gtex_protein_table}
\title{GTEx protein tissue specificity annotations}
\format{
A tibble object with 394,400 rows and 3 columns:
\describe{
  \item{tissue}{<chr> Tissue name}
  \item{gene}{<chr> HGNC gene symbol}
  \item{significant}{<lgl> whether or not the gene is tissue-enriched} 
}
}
\source{
Table S4 of Jiang et al. downloaded on 2023-10-13 from:
\url{https://www.sciencedirect.com/science/article/pii/S0092867420310783}
}
\usage{
gtex_protein_table
}
\description{
A dataset containing annotations for 12,325 tissue-enriched genes in 32 tissues derived from
GTEx proteomics data. See references for definition of tissue-enriched genes.
}
\references{
Jiang L, Wang M, Lin S, et al. A Quantitative Proteome Map of the Human Body.
Cell. 2020;183(1):269-283.e19. doi:10.1016/j.cell.2020.08.036
}
\seealso{
Other datasets: 
\code{\link{accession_gene_table}},
\code{\link{bioplex_table}},
\code{\link{example_data2}},
\code{\link{example_data3}},
\code{\link{example_data}},
\code{\link{genes_snps_hash}},
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gtex_brain_table}},
\code{\link{gtex_rna_table}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{hpa_rna_table}},
\code{\link{inweb_table}},
\code{\link{irefindex_table}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c2_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_c8_table}},
\code{\link{msigdb_h_table}},
\code{\link{string_table}}
}
\concept{datasets}
\keyword{datasets}
