% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GTEX_table}
\alias{GTEX_table}
\title{Genotype tissue expression (GTEx)}
\format{data.frame object with: tissue, gene, significant (indicating whether the gene is elvated in the tissue.)}
\source{
downloaded on 2020-06-15 (\url{https://data.broadinstitute.org/alkesgroup/LDSCORE/LDSC_SEG_ldscores/tstats/}).
}
\usage{
GTEX_table
}
\description{
A table containing classification of all human protein-coding genes based on transcript expression levels in 53 tissues
(20,025 unique genes). Constructed from a gene by tissue matrix containing t-statistics from Finucane et al (see reference and source). 
Genes are ranked by the t-statistic, and the top 10% of genes in a tissue are considered 'elevated'.
}
\references{
Finucane, H.K., Reshef, Y.A., Anttila, V. et al. Heritability enrichment of specifically expressed genes identifies disease-relevant tissues and cell types. Nat Genet 50, 621–629 (2018).
}
\seealso{
Other datasets: 
\code{\link{accession_gene_table}},
\code{\link{example_data2}},
\code{\link{example_data3}},
\code{\link{example_data}},
\code{\link{genes_snps}},
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{hpa_table}},
\code{\link{inweb_hash}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c2_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_h_table}}
}
\concept{datasets}
\keyword{datasets}
