% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hgnc_group_table}
\alias{hgnc_group_table}
\title{HGNC gene groups}
\format{
A tibble object with 30,791 rows and 3 columns:
\describe{
  \item{Gene.symbol}{<chr> HGNC gene symbol}
  \item{Group.ID}{<int> Gene group ID}
  \item{Group.name}{<chr> Gene group name}
}
}
\source{
HGNC genes within groups dataset downloaded on 2023-10-02 from:
\url{https://www.genenames.org/download/statistics-and-files/}
}
\usage{
hgnc_group_table
}
\description{
A dataset containing annotations for 25,637 genes in 1,622 HGNC gene groups.
}
\references{
Seal RL, Braschi B, Gray K, et al. Genenames.org: the HGNC resources in 2023.
Nucleic Acids Res. 2023;51(D1):D1003-D1009. doi:10.1093/nar/gkac888
}
\seealso{
Other datasets: 
\code{\link{accession_gene_table}},
\code{\link{bioplex_table}},
\code{\link{example_data2}},
\code{\link{example_data3}},
\code{\link{example_data}},
\code{\link{genes_snps}},
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gtex_protein}},
\code{\link{gtex_rna}},
\code{\link{gwas_table}},
\code{\link{hpa_rna}},
\code{\link{inweb_table}},
\code{\link{irefindex_table}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c2_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_c8_table}},
\code{\link{msigdb_h_table}}
}
\concept{datasets}
\keyword{datasets}
