% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hpa_rna}
\alias{hpa_rna}
\title{Human Protein Atlas (HPA) RNA tissue specificity annotations}
\format{
A data frame with 571,504 rows and 3 variables:
\describe{
  \item{tissue}{Tissue name}
  \item{gene}{HGNC gene symbol}
  \item{significant}{Logical variable: "TRUE" = gene has elevated expression in tissue;
                     "FALSE" = gene is detected (but not elevated) in tissue} 
}
}
\source{
proteinatlas.tsv.zip datasets (v19.3) downloaded on 2020-06-12 from:
\url{https://www.proteinatlas.org/about/download}
}
\usage{
hpa_rna
}
\description{
A dataset containing tissue specificity annotations for 19,634 genes in 33 tissues derived from 
HPA RNA-sequencing data. See cited reference and \url{https://www.proteinatlas.org/humanproteome/tissue/tissue+specific}
for details on definition of tissue-specific (i.e. tissue-elevated) genes.
}
\references{
Uhlén M, Fagerberg L, Hallström BM, et al. Proteomics. Tissue-based map of the human proteome.
Science. 2015;347(6220):1260419. doi:10.1126/science.1260419
}
\seealso{
Other datasets: 
\code{\link{accession_gene_table}},
\code{\link{bioplex_table}},
\code{\link{example_data2}},
\code{\link{example_data3}},
\code{\link{example_data}},
\code{\link{genes_snps}},
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gtex_protein}},
\code{\link{gtex_rna}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{inweb_table}},
\code{\link{irefindex_table}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c2_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_h_table}}
}
\concept{datasets}
\keyword{datasets}
