% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hpa_rna_table}
\alias{hpa_rna_table}
\title{Human Protein Atlas (HPA) RNA tissue specificity annotations}
\format{
A tibble object with 688,104 rows and 3 columns:
\describe{
  \item{tissue}{<chr> Tissue name}
  \item{gene}{<chr> HGNC gene symbol}
  \item{significant}{<lgl> whether or not gene is tissue-elevated}
}
}
\source{
Human Protein Atlas data (v23.0) downloaded from: \url{https://www.proteinatlas.org/about/download}
}
\usage{
hpa_rna_table
}
\description{
A dataset containing annotations for 19,114 tissue-elevated genes in 36 tissues derived from 
HPA consensus transcriptomics data. See references for definition of tissue-elevated genes.
}
\references{
Uhlén M, Fagerberg L, Hallström BM, et al. Proteomics. Tissue-based map of the human proteome.
Science. 2015;347(6220):1260419. doi:10.1126/science.1260419 

See additional documentation at: https://www.proteinatlas.org/humanproteome/tissue/tissue+specific
}
\seealso{
Other datasets: 
\code{\link{accession_gene_table}},
\code{\link{bioplex_table}},
\code{\link{example_data2}},
\code{\link{example_data3}},
\code{\link{example_data}},
\code{\link{genes_snps_hash}},
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gtex_brain_table}},
\code{\link{gtex_protein_table}},
\code{\link{gtex_rna_table}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{inweb_table}},
\code{\link{irefindex_table}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c2_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_c8_table}},
\code{\link{msigdb_h_table}},
\code{\link{string_table}}
}
\concept{datasets}
\keyword{datasets}
