% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hpa_table}
\alias{hpa_table}
\title{Subsetted Human Protein Atlas annotation table}
\format{data.frame object}
\source{
merge between proteinatlas.tsv.zip and normal_tissue.tsv.zip downloaded on 2020-06-12 (https://www.proteinatlas.org/about/download).
}
\usage{
hpa_table
}
\description{
A table containing classification of all human protein-coding genes based on transcript expression levels in 32 tissues.
}
\seealso{
Other datasets: 
\code{\link{accession_gene_table}},
\code{\link{example_data2}},
\code{\link{example_data3}},
\code{\link{example_data}},
\code{\link{genes_snps}},
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{inweb_hash}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c2_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_h_table}}
}
\concept{datasets}
\keyword{datasets}
