% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_translate_significance_thresholds.R
\name{html_translate_significance_tresholds}
\alias{html_translate_significance_tresholds}
\title{translate significance and logfc thresholds to html text}
\usage{
html_translate_significance_tresholds(
  fc,
  fc_dir,
  sig_type,
  fdr_threshs,
  pval_thresh
)
}
\arguments{
\item{fc}{the logFC thresholds, i.e. a numeric value.}

\item{fc_dir}{the direction of the logFC: negative, positive or both.}

\item{sig_type}{the significance type: fdr or pvalue.}

\item{pval_thresh}{the threshold for the p-value.}

\item{fdr_thresh}{the fdr fdr thresholds}
}
\description{
translate significance and logfc thresholds to html that
can be used in the shiny app for dynamic ui.
}
