% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hyperlink}
\alias{hyperlink}
\title{hyperlink}
\usage{
hyperlink(url, text)
}
\arguments{
\item{url}{what is the url?}

\item{text}{what should be displayed? Default is NULL,
which just displays the URL.}
}
\description{
convert a string into a hyperlink
}
\concept{html}
