% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_enriched_proteins.R
\name{id_enriched_proteins}
\alias{id_enriched_proteins}
\title{Identify enriched proteins}
\usage{
id_enriched_proteins(
  df,
  logfc_dir = "positive",
  logfc_cutoff = NULL,
  p_cutoff = NULL,
  fdr_cutoff = 0.1
)
}
\arguments{
\item{df}{data.frame contaiing logFC, pvalue, and FDR columns}

\item{logfc_dir}{string indicating direction of logFC cutoff ("positive" or "negative")}

\item{logfc_cutoff}{numeric value indicating logFC cutoff}

\item{p_cutoff}{numeric value indicating pvalue cutoff}

\item{fdr_cutoff}{numeric value indicating FDR cutoff}
}
\value{
data.frame containing df + signficanct column of boolean values indicating which df rows (proteins) are enriched
}
\description{
Identify enriched proteins based in input thresholds.
}
