% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_na.R
\name{impute_na}
\alias{impute_na}
\title{Impute missing values in proteomics data}
\usage{
impute_na(df, shift = 1.8, width = 0.3)
}
\arguments{
\item{df}{data.frame containing columns with numeric values (e.g. normalized,
log2-transformed protein intensities) and NAs}

\item{shift}{numeric value}

\item{width}{numeric value}
}
\value{
data.frame containing same columns as input df (with NAs replaceD), and
corresponding "_imp" columns indicating if values were imputed (TRUE) vs. observed (FALSE).
}
\description{
replace missing values (NAs) in each sample (i.e. column) by
sampling from a normal distribution with mean of x - y*shift and SD of y*width,
where x and y are the mean and SD of the observed values in the sample.
}
