% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_format.R
\name{input_format}
\alias{input_format}
\title{check input format}
\usage{
input_format(tabl)
}
\arguments{
\item{tabl}{a data.frame}
}
\description{
checks input format.
}
\note{
an item in the approved format list
can appear more than once.
}
\examples{
\dontrun{
tabl1 = data.frame(gene = letters,rep1=runif(26), rep2=rnorm(26))
tabl2 = data.frame(gene = letters,rep1=runif(26), rep2=rnorm(26), somethingelse=1:26)
input_format(tabl1)
}
}
