% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inweb_table}
\alias{inweb_table}
\title{InWeb protein-protein interactions}
\format{
A data frame with 883,356 rows and 4 variables:
\describe{
  \item{Gene1}{HGNC gene symbol of first protein}
  \item{Gene2}{HGNC gene symbol of second protein}
  \item{Score}{Interaction confidence score}
  \item{Source}{InWeb source version, "InWeb_IM" or "InWeb3"}
}
}
\source{
See cited references; InWeb_IM data (2016-9-12 release) downloaded from:
\url{https://inbio-discover.intomics.com/map.html#downloads}
}
\usage{
inweb_table
}
\description{
A dataset containing published protein-protein interactions for 18,814 human genes in
InWeb_InBioMap (InWeb_IM) or InWeb3.
}
\references{
Li T, Wernersson R, Hansen RB, et al. A scored human protein-protein interaction network to catalyze
genomic interpretation. Nat Methods. 2017;14(1):61-64. doi:10.1038/nmeth.4083  

Lage K, Karlberg EO, Størling ZM, et al. A human phenome-interactome network of protein complexes implicated
in genetic disorders. Nat Biotechnol. 2007;25(3):309-316. doi:10.1038/nbt1295
}
\seealso{
Other datasets: 
\code{\link{accession_gene_table}},
\code{\link{example_data2}},
\code{\link{example_data3}},
\code{\link{example_data}},
\code{\link{genes_snps}},
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gtex_proteome_table}},
\code{\link{gtex_table}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{hpa_table}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c2_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_h_table}}
}
\concept{datasets}
\keyword{datasets}
