% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{irefindex_table}
\alias{irefindex_table}
\title{IRefIndex 17.0 interaction data.base}
\format{
A data frame with 406,902 rows and 5 variables:
\describe{
  \item{Gene1}{HGNC gene symbol of first protein (bait/prey)}
  \item{Gene2}{HGNC gene symbol of second protein (bait/prey)}
  \item{Score.hpr.max}{Highest PubMed Identifier (PMID) Reuse.}
  \item{Score.lpr.max}{Lowest PubMed Identifier (PMID) Reuse.}
  \item{Score.np.max}{Number of publications.}
}
}
\source{
iRefIndex 17.0 (9606.mitab.27062020.txt) downloaded on Jan-06-2021 
from \url{https://irefindex.vib.be/download/irefindex/data/archive/release_17.0/psi_mitab/MITAB2.6/}
}
\usage{
irefindex_table
}
\description{
A dataset containing 406,902 protein-protein interactions for human 17,437 genes.
}
\details{
Understanding score.x.max: Since a binary interaction may appear multiple times in IRefIndex, 
each score entry: 'hpr', 'lpr' and 'np' were aggregated and the maximum values 
were kept.
}
\references{

}
\seealso{
Other datasets: 
\code{\link{accession_gene_table}},
\code{\link{bioplex_table}},
\code{\link{example_data2}},
\code{\link{example_data3}},
\code{\link{example_data}},
\code{\link{genes_snps}},
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gtex_protein}},
\code{\link{gtex_rna}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{hpa_rna}},
\code{\link{inweb_table}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c2_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_h_table}}
}
\concept{datasets}
\keyword{datasets}
