% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{italics}
\alias{italics}
\title{italics}
\usage{
italics(x)
}
\arguments{
\item{x}{string}
}
\description{
make text html italics
}
\seealso{
Other misc: 
\code{\link{\%nin\%}()},
\code{\link{as.bait}()},
\code{\link{bold}()},
\code{\link{catf}()},
\code{\link{color_distinct}()},
\code{\link{color_gradient}()},
\code{\link{enumerate_replicate_combinations}()},
\code{\link{lun}()},
\code{\link{null_omit}()}
}
\concept{misc}
