% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapply_calc_hyper.R
\name{lapply_calc_hyper}
\alias{lapply_calc_hyper}
\title{lapply calc_hyper to multiple datasets}
\usage{
lapply_calc_hyper(
  data,
  reference,
  col.by = "tissue",
  bait = NULL,
  p.adj.method = "fdr",
  intersectN = T,
  verbose = F
)
}
\arguments{
\item{data}{proteomic data with gene and significant columns.}

\item{reference}{table with col.by, gene, significant in columns.}

\item{col.by}{string. What column contains the group?}

\item{bait}{bait. String. Passed to \code{?calc_hyper}.}

\item{p.adj.method}{Default is fdr. See p.adjust.methods in \code{?stats::p.adjust}.}

\item{intersectN}{Boolean indicating if the total population should be the intersect
of the two datasets, when calculating hypergeometric overlap.}

\item{verbose}{print out progress}
}
\description{
calculate hypergeometric enrichment of genes in tissue and 
subsequent adjusting for multiple testing.
}
\examples{
\dontrun{
# check for enrichment in GTEx
data(gtex_table)
data(example_data)
data = example_data \%>\% calc_mod_ttest \%>\% id_significant_proteins()
gtex_enrichment = lapply_calc_hyper(data, gtex_table)
gtex_enrichment
}
}
