% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_genoppi.R
\name{launch_genoppi}
\alias{launch_genoppi}
\title{launch genoppi}
\usage{
launch_genoppi(local = T)
}
\arguments{
\item{local}{boolean. Launch locally on your machine or try connect to server (to be implemented)}
}
\description{
launches genoppi in a seperte shiny application
}
