% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overlay.R
\name{list_to_df}
\alias{list_to_df}
\title{concert genoppi genelist to data.frame}
\usage{
list_to_df(lst)
}
\arguments{
\item{lst}{a names reference list of data.frames}
}
\value{
a data.frame
}
\description{
converts a named list of datasets to
a single data.frame that also contains shape, label and color.
}
\note{
internal
}
