% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_interactive.R
\name{make_interactive}
\alias{make_interactive}
\title{Make ggplot interactive using plotly}
\usage{
make_interactive(p, source = NULL, legend = T, sig_text = "")
}
\arguments{
\item{p}{a ggplot (See details)}

\item{source}{string. Used to connect plotly object other plotly objects.}

\item{legend}{boolean. Show legend for significant interactors?}

\item{sig_text}{what legend text should be added to a significant item?}
}
\description{
Convert a volcano or scatter plot to an interactive plotly object.
}
\details{
In order to make a plot interactive, they must be generated using
\code{plot_volcano_basic}, \code{plot_scatter_basic} or/and \code{plot_overlay}.
}
\examples{
\dontrun{
df <- example_data \%>\%
  calc_mod_ttest() \%>\%
  id_significant_proteins()

# overlay simple a bait and make interactive
plt <- df \%>\% 
  plot_volcano_basic() \%>\%
  plot_overlay(as.bait('BCL2')) \%>\%
 volcano_theme() \%>\% 
  make_interactive()

plt

# save as html widget (requires htmlwidgets package)
htmlwidgets::saveWidget(as_widget(plt), "index.html")
}
}
\seealso{
Other interactivity: 
\code{\link{set_names_by_dataset}()},
\code{\link{tabulate_markers}()}
}
\concept{interactivity}
