% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{lun}
\alias{lun}
\title{misc tools}
\usage{
lun(x)
}
\arguments{
\item{x}{a vector or list of items.}
}
\description{
length of unique items.
}
\seealso{
Other misc: 
\code{\link{\%nin\%}()},
\code{\link{as.bait}()},
\code{\link{bold}()},
\code{\link{catf}()},
\code{\link{color_distinct}()},
\code{\link{color_gradient}()},
\code{\link{enumerate_replicate_combinations}()},
\code{\link{italics}()},
\code{\link{null_omit}()}
}
\concept{misc}
