% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_ggplot_from_overlay.R
\name{modify_ggplot_from_overlay}
\alias{modify_ggplot_from_overlay}
\title{modify ggplot from overlay}
\usage{
modify_ggplot_from_overlay(p)
}
\arguments{
\item{p}{a ggplot object also containing a data.frame \code{overlay}.}
}
\description{
a function used to modify a ggplot, e.g. a color
from an overlayed genelist. This is primarily used by shiny/plotly.
}
\note{
currently only handles color.
}
