% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msigdb_c2_table}
\alias{msigdb_c2_table}
\title{MSigDB C2 collection (curated gene sets) annotation table}
\format{data.frame object with: Gene.symbol, Set.name}
\source{
c2.all.v7.1.symbols.gmt file downloaded on 2020-04-03 (https://www.gsea-msigdb.org/gsea/msigdb/collections.jsp#H).
}
\usage{
msigdb_c2_table
}
\description{
A table containing 479,444 MSigDB curated gene set annotations (20,738 unique gene symbols, 5,529 unique gene sets).
}
\seealso{
Other datasets: 
\code{\link{accession_gene_table}},
\code{\link{example_data}},
\code{\link{genes_snps}},
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{inweb_hash}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_h_table}}
}
\concept{datasets}
\keyword{datasets}
