% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msigdb_c8_table}
\alias{msigdb_c8_table}
\title{MSigDB C8 collection (cell type signature gene sets)}
\format{
A tibble object with 149,795 rows and 2 columns:
\describe{
  \item{Gene.symbol}{<chr> HGNC gene symbol}
  \item{Set.name}{<chr> Gene set name}
}
}
\source{
MSigDB Human Gene Set GMT file set (v2023.1.Hs release) downloaded from:
\url{https://www.gsea-msigdb.org/gsea/downloads.jsp}
}
\usage{
msigdb_c8_table
}
\description{
A dataset containing annotations for 20,393 genes in 830 MSigDB cell type signature gene sets.
}
\references{
Liberzon A, Birger C, Thorvaldsdóttir H, Ghandi M, Mesirov JP, Tamayo P.
The Molecular Signatures Database (MSigDB) hallmark gene set collection.
Cell Syst. 2015;1(6):417-425. doi:10.1016/j.cels.2015.12.004 

See additional references at: https://www.gsea-msigdb.org/gsea/msigdb/human/collection_details.jsp
}
\seealso{
Other datasets: 
\code{\link{accession_gene_table}},
\code{\link{bioplex_table}},
\code{\link{example_data2}},
\code{\link{example_data3}},
\code{\link{example_data}},
\code{\link{genes_snps}},
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gtex_protein}},
\code{\link{gtex_rna}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{hpa_rna}},
\code{\link{inweb_table}},
\code{\link{irefindex_table}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c2_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_h_table}}
}
\concept{datasets}
\keyword{datasets}
