% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdb_table_data.R
\docType{data}
\name{msigdb_table}
\alias{msigdb_table}
\title{MSigDB gene set annotation table}
\format{data.frame object with: Gene.symbol, Set.name}
\source{
"msigdb.v7.0.symbols.gmt. Downloaded on 2020-03-30 (https://www.gsea-msigdb.org/gsea/downloads.jsp#msigdb).
}
\usage{
msigdb_table
}
\description{
A table containing 2,639,029 MSigDB gene set annotations (38,040 unique gene symbols, 22,596 unique GO MF terms).
}
\seealso{
Other datasets: 
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}}
}
\concept{datasets}
\keyword{datasets}
