% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{null_omit}
\alias{null_omit}
\title{omit nulls from list}
\usage{
null_omit(lst)
}
\arguments{
\item{lst}{an R list}
}
\description{
remove NULLs in list
}
\seealso{
Other misc: 
\code{\link{\%nin\%}()},
\code{\link{as.bait}()},
\code{\link{as.goi}()},
\code{\link{bold}()},
\code{\link{catf}()},
\code{\link{color_distinct}()},
\code{\link{color_gradient}()},
\code{\link{enumerate_replicate_combinations}()},
\code{\link{get_gg_legend}()},
\code{\link{hline}()},
\code{\link{italics}()},
\code{\link{lun}()},
\code{\link{shape_to_symbol}()},
\code{\link{symbol_to_shape}()},
\code{\link{vline}()}
}
\concept{misc}
