% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rd.R
\name{parse_rd}
\alias{parse_rd}
\title{parse documentation files for content}
\usage{
parse_rd(
  content,
  what = c("description", "title", "format", "references"),
  raw = F,
  combine = T,
  combine.with = "\\n"
)
}
\arguments{
\item{content}{a vector of strings.}

\item{what}{a vector of search terms.}

\item{raw}{boolean. Should the raw data be returned?}

\item{combine}{boolean. Should the vectors in the list be combined?}

\item{combine.with}{string. What should the vectors be collapsed with.}
}
\description{
A function that parsed markdown documentation files
that has been written in \code{devtools::document}.
}
\examples{
\dontrun{
content = readLines('man/GTEX_table.Rd')
parse_rd(content)
}
}
