% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_uploaded_file.R
\name{parse_uploaded_file}
\alias{parse_uploaded_file}
\title{parse uploaded file}
\usage{
parse_uploaded_file(path)
}
\arguments{
\item{data}{a string, data path}
}
\value{
a data.frame ready for further analysis in genoppi
}
\description{
the following actions are performed sequentially
1) call read_input to get the file, 2) check whether accession_number
column is present, and if so, map to gene id; 3) calculate moderated
t-test on the data
}
\seealso{
Other shiny: 
\code{\link{add_hover_lines_volcano}()},
\code{\link{add_layout_html_axes_scatterplot}()},
\code{\link{add_layout_html_axes_volcano}()},
\code{\link{add_markers_search_pathway}()},
\code{\link{add_markers_search}()},
\code{\link{calc_logfc_limit}()}
}
\concept{shiny}
