% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcnet_table}
\alias{pcnet_table}
\title{PCNet (Parsimonious Composite Network)}
\format{
TO ADD
\describe{
  \item{Gene1}{HGNC gene symbol of first protein}
  \item{Gene2}{HGNC gene symbol of second protein}
}
}
\source{
PCNet (Parsimonious Composite Network) data downloaded on 2023-10-19 from:
\url{https://www.ndexbio.org/v2/network/f93f402c-86d4-11e7-a10d-0ac135e8bacf?download=true}
}
\usage{
pcnet_table
}
\description{
TO ADD
}
\details{
TO ADD
}
\references{
Huang JK, Carlin DE, Yu MK, Zhang W, Kreisberg JF, Tamayo P, Ideker T. Systematic Evaluation of Molecular Networks
for Discovery of Disease Genes. Cell Syst. 2018 Apr 25;6(4):484-495.e5. doi: 10.1016/j.cels.2018.03.001. Epub 2018
Mar 28. PMID: 29605183; PMCID: PMC5920724.
}
\seealso{
Other datasets: 
\code{\link{accession_gene_table}},
\code{\link{bioplex_table}},
\code{\link{example_data2}},
\code{\link{example_data3}},
\code{\link{example_data}},
\code{\link{genes_snps_hash}},
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gtex_brain_table}},
\code{\link{gtex_protein_table}},
\code{\link{gtex_rna_table}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{hpa_rna_table}},
\code{\link{huri_table}},
\code{\link{inweb_table}},
\code{\link{irefindex_table}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c2_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_c8_table}},
\code{\link{msigdb_h_table}},
\code{\link{string_table}}
}
\concept{datasets}
\keyword{datasets}
