% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overlay.R
\name{plot_overlay}
\alias{plot_overlay}
\title{superimpose genesets onto plots}
\usage{
plot_overlay(
  p,
  reference,
  match = "gene",
  size_gg = 3.5,
  stroke = 0.75,
  sig_text = NULL,
  insig_text = NULL,
  label = NULL,
  label_size = NULL,
  label_color = "black",
  label_box_padding = 0.3,
  label_point_padding = 0.5,
  label_arrowhead_size = 0.01,
  legend_nchar_max = NULL,
  nchar_max_collapse = "<br>"
)
}
\arguments{
\item{p}{A ggplot object. Usually passed down from \code{plot_volcano_basic} or \code{plot_scatter_basic}.}

\item{reference}{a list of data.frames that are preferably named. The name of the list will passed down to
the data.frame as the column 'dataset'. Alternatively, the dataset can have a column name dataset.}

\item{match}{by what string should the ggplot and overlay be merged? Default is 'gene'.}

\item{size_gg}{size of ggplot points.}

\item{stroke}{numeric. The width of the outline/borders.}

\item{sig_text}{string. text for significant interactors to be displayed in legend.}

\item{insig_text}{string. Text for non-significant interactors to be displayed in legend.}

\item{label}{A boolean. This will overwrite the \code{label} column in the reference data.frame.}

\item{label_size}{numeric. Size of label. This will overwrite the \code{label.size} column in the reference data.frame.}

\item{label_color}{the color of the label. Default is black.}

\item{label_box_padding}{Amount of padding around bounding box. See \code{?ggrepel::geom_text_repel} 
for more details.}

\item{label_point_padding}{Amount of padding around label. See \code{?ggrepel::geom_text_repel}.}

\item{label_arrowhead_size}{The size of the arrowhead. 0 means no arrowhead.}

\item{legend_nchar_max}{maximum amount of allowed characters in the legend.}

\item{nchar_max_collapse}{what charcter should be used for line break? Default is HTML line break \code{"<br>".}}
}
\value{
a ggplot
}
\description{
Takes a ggplot2 object and overlays genes from a reference data.frame. This data.frame
can be constructed manually or generated with a variety of functions, e.g. \code{get_inweb_list},
\code{get_irefindex_list} or \code{get_tissue_list}.

  
The function uses the mapping and plot environment from a previous ggplot to add an overlay
with a matching 'reference' data.frame. This yields a new plot with items that intersect the 
the original ggplot data and the reference data. Can be applied iteratively. 
The user can input certain columns in the reference to gain more control over the plotting 
parameters of individual points. See notes for additional details.
}
\note{
The following data.frame headers for a reference are accepted:
\itemize{
 \item{"gene"}{ A string tht indicates the gene name.}
 \item{"label"} {A boolen that indicates whether the label should be plotted.}
 \item{"label.size"}{A numeric that indicates the size of the label items.}
}
See ?validate_reference for additional details.
}
\examples{
\dontrun{
df <- example_data \%>\%
  calc_mod_ttest() \%>\%
  id_significant_proteins()

# overlay simple  a bait
df \%>\% 
  plot_volcano_basic() \%>\%
  plot_overlay(as.bait('BCL2')) \%>\%
  volcano_theme()

# make a custom overlay with custom colors
myoverlay = data.frame(gene = c('FUS', 'RBMX'),
                       col_significant = c('cyan', 'blue'),
                       col_other = c('grey', 'grey'), 
                       dataset = c('group 1', 'group 2'))

# plot overlay
df \%>\% 
  plot_volcano_basic() \%>\%
  plot_overlay(list(overlay = myoverlay)) \%>\%
  volcano_theme() 

# plot multiple overlays
df \%>\% 
  plot_volcano_basic() \%>\%
  plot_overlay(list(overlay = myoverlay)) \%>\%
  plot_overlay(as.bait('BCL2')) \%>\%
  volcano_theme() 
}
}
