% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_replicate_ggpair.R
\name{plot_replicate_ggpair}
\alias{plot_replicate_ggpair}
\title{plot ggpair plot for given columns from the given data frame}
\usage{
plot_replicate_ggpair(data, columns, overlay_sigificance = FALSE, title = "")
}
\arguments{
\item{data}{data frame containing data to be plotted}

\item{columns}{vector containing column names of the data frame provided}

\item{overlay_sigificance}{Boolean indicating whether the sigificant proteins should be overlaid with green color dots (can only be used when a Boolean significant column is present).}

\item{title}{optional title for the ggpair plot}
}
\value{
a ggpair plot showing correlation between samples, controls, sample and contorl, or replicate logFC.
}
\description{
plot ggpair plot for given columns from the given data frame
}
\examples{
\dontrun{
p = example_data \%>\% 
  calc_mod_ttest() \%>\%
  id_significant_proteins(fdr_cutoff = 0.1) \%>\%
  plot_replicate_ggpair()
}
}
