% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter_basic.R
\name{plot_scatter_basic}
\alias{plot_scatter_basic}
\title{plot a basic scatter plot of replicates}
\usage{
plot_scatter_basic(
  df,
  repA = "rep1",
  repB = "rep2",
  size_gg = 3,
  col_significant = "#41AB5D",
  col_other = "grey",
  sig_text = "(significant)",
  insig_text = "(not significant)",
  shape = 21,
  stroke = 0.2,
  col_border = NULL,
  unit = log[2] ~ "[Fold Change]"
)
}
\arguments{
\item{df}{data.frame with at least columns \code{gene}, \code{significant}
and some \code{rep[0-9]}.}

\item{repA}{string that is column in df. Expects name to be in the form of
rep[0-9], e.g. 'rep1'.}

\item{repB}{string that is column in df. Expects name to be in the form of
rep[0-9], e.g. 'rep1'.}

\item{size_gg}{numeric. Size of point.}

\item{col_significant}{string. Color for significant protein interactors.}

\item{col_other}{string. Color for other protein interactors.}

\item{sig_text}{string. Text for significant interactor to be displayed in
legend.}

\item{insig_text}{string. Text for non-significant interactors to be
displayed in legend.}

\item{shape}{integer. The shape of the point. Default is 21 (circle).}

\item{stroke}{numeric. The stroke width.}

\item{col_border}{string. The color of the borders/outline.}

\item{unit}{unit displayed in the x and y axis}
}
\description{
Plot protein-level data in a sactter plot.
}
\examples{
\dontrun{
# run example data
p = example_data \%>\% 
calc_mod_ttest() \%>\%
  id_significant_proteins(fdr_cutoff = 0.1) \%>\%
  plot_scatter_basic() \%>\%
  plot_overlay(as.bait('BCL2'))
  
p + ggtitle('Example scatter plot')
}
}
