% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter_basic.R
\name{plot_scatter_basic_all}
\alias{plot_scatter_basic_all}
\title{plot a list of basic scatter plots}
\usage{
plot_scatter_basic_all(
  df,
  size_point = 3,
  col_significant = "#41AB5D",
  col_other = "grey"
)
}
\arguments{
\item{df}{A data.frame containing at least gene, significant, replicate/triplicate columns.}

\item{size_point}{size of point.}

\item{col_significant}{color for significant protein interactors.}

\item{col_other}{color for other protein interactors.}
}
\value{
a list of gg scatter plots.
}
\description{
Draw qll pairs of replicates in multiple scatter plot.
}
