% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tissue_enrichment.R
\name{plot_tissue_enrichment}
\alias{plot_tissue_enrichment}
\title{plot tissue enrichment}
\usage{
plot_tissue_enrichment(
  data,
  col.tissue,
  col.value,
  xlab = "tissue",
  ylab = "-log10(Hypergeometric P-value)",
  pvalue.line = NULL
)
}
\arguments{
\item{data}{a data.frame}

\item{col.tissue}{string. Column containing tissue names.}

\item{col.value}{string. Column containing pvalue, FDR, enrichment for plotting.}

\item{xlab}{x label}

\item{ylab}{y label}

\item{pvalue.line}{pvalue threshold to be displayed via a horisontal line.}
}
\description{
constructs a bar chart using ggplot to plot tissue enrichment.
}
\seealso{
Other ggplot: 
\code{\link{scatter_theme}()}
}
\concept{ggplot}
