% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_genoppi_venn.R
\name{plot_venn}
\alias{plot_venn}
\title{plot venn diagram}
\usage{
plot_venn(v, scale = 0.9)
}
\arguments{
\item{v}{a venn diagram made from package 'VennDiagram' or from genoppi::draw_genoppi_venn().}

\item{scale}{what scale should it be plotted at?}
}
\description{
plots a venn at a reduced scale to allow for longer
diagram names.
}
