% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Plot basic volcano}
\usage{
takes in a data.frane with the columns gene, logFC, pvalue and significant 
to draw a volcano. Optionally, a column indicating 'color' (string) can be supplied to 
indicate the volcano color scheme and whether to draw names of specific proteins.
}
\arguments{
\item{df}{a data.frame with at least columns gene, logFC, pvalue and significant.}

\item{bait}{the gene name of the bait.}

\item{col_other}{the color of non-significnt proteins/rows.}

\item{col_significant}{the color of significant proteins/rows.}
}
\description{
A function that draw a basic volcano plot.
}
