% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano_basic.R
\name{plot_volcano_basic}
\alias{plot_volcano_basic}
\title{Plot basic volcano}
\usage{
plot_volcano_basic(
  df,
  col_significant = "#41AB5D",
  col_other = "grey",
  sig_text = "(significant)",
  insig_text = "(not significant)",
  size_gg = 3,
  shape = 21,
  stroke = 0.2,
  col_border = NULL,
  plot_segments = F
)
}
\arguments{
\item{df}{data.frame with at least columns \code{gene}, \code{logFC}, \code{pvalue} and \code{significant}.}

\item{col_significant}{string. The color of significant proteins/genes.}

\item{col_other}{string. The color of non-significant proteins/genes.}

\item{sig_text}{string. Text for significant interactor to be displayed in legend.}

\item{insig_text}{string. Text for non-significant interactors to be displayed in legend.}

\item{size_gg}{numeric. The size of the points.}

\item{shape}{integer. The shape of the points. Default to 21 (circles).}

\item{stroke}{numeric. Stroke width.}

\item{col_border}{string. A color for the outline of points.}

\item{plot_segments}{boolean. Plot volcano plot axis segments, i.e. two lines with x and y axis intercept.}
}
\description{
Plot protein-level data in a basic volcano plot. Can be chained with \code{plot_overlay} to produce overlays.
}
\examples{
\dontrun{
p = example_data \%>\% 
calc_mod_ttest() \%>\%
  id_significant_proteins(fdr_cutoff = 0.1) \%>\%
  plot_volcano_basic() \%>\%
  plot_overlay(as.bait('BCL2')) 
p + ggtitle('Example volcano plot') + theme_classic()

}
}
