% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tissue_enrichment.R
\name{plotly_tissue_enrichment}
\alias{plotly_tissue_enrichment}
\title{plot tissue enrichment}
\usage{
plotly_tissue_enrichment(
  data,
  col.tissue,
  col.value,
  col.value.text = NULL,
  pvalue.line = NULL,
  xlab = "",
  ylab = "",
  title = "",
  xlim = NULL,
  color.significant = "red",
  color.other = "orange",
  col.value.order = NULL
)
}
\arguments{
\item{data}{a data.frame}

\item{col.tissue}{string. Column containing tissue names.}

\item{col.value}{string. Column containing pvalue, FDR, enrichment for plotting.}

\item{col.value.text}{string. Column containing text that should be displayed when hovering over bar.}

\item{pvalue.line}{p-value horizontal line.}

\item{xlab}{string. x label.}

\item{ylab}{string. y label.}

\item{title}{string. title of plot.}

\item{xlim}{vector of length 2.}

\item{color.significant}{color of significant values. default is red.}

\item{color.other}{color of non-signfificant values. Default is orange.}

\item{col.value.order}{string that is either 'increasing' or 'decreasing'. Orders barplot accordingly.}
}
\description{
constructs a bar chart using ggplot to plot tissue enrichment.
}
\seealso{
Other plotly: 
\code{\link{add_plotly_layout_scatter}()},
\code{\link{add_plotly_layout_volcano}()},
\code{\link{add_plotly_line_unity}()},
\code{\link{add_plotly_threshold_lines}()},
\code{\link{add_plotly_trace}()}
}
\concept{plotly}
