% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_input.R
\name{read_input}
\alias{read_input}
\title{Read in proteomic data from file}
\usage{
read_input(filename, sep = "")
}
\arguments{
\item{filename}{path to input file.}

\item{sep}{the field character seperator, see \code{?read.table}.}
}
\value{
list contaiing two objects, data and format. data: input data.frame. format: list returned by check_input(), see \code{?check_input}.
}
\description{
Read in file containing proteomic data as data.frame. Call check_input() to check data format.
}
\examples{
\dontrun{
df1 <- data.frame(gene=letters, rep1=rnorm(26), rep2=rnorm(26), rep3=rnorm(26))
tmp <- tempfile()
write.table(df1, tmp, quote=F, sep="\t", row.names=F)
input <- read_input(tmp, sep="\t")
} 

}
