% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snp_list.R
\name{read_snp_list}
\alias{read_snp_list}
\title{Read in SNP list data}
\usage{
read_snp_list(infile, genes)
}
\arguments{
\item{infile}{a.data.frame or a file path of SNP list file (one column of SNP IDs with no header)}

\item{genes}{vector of gene names (genes detected in proteomic data)}
}
\value{
data.frame containing gene and SNP columns (restricting to genes in input genes vector; genes and SNPs can appear multiple times)
}
\description{
Create gene list data.frame from input SNP list file
}
