% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.R
\docType{data}
\name{ref1}
\alias{ref1}
\title{plot_overlay}
\format{An object of class \code{data.frame} with 1 rows and 5 columns.}
\usage{
ref1
}
\description{
Overlays a ggplot with a reference data.set, e.g. a bait or a genelist
}
\keyword{datasets}
