% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_genoppi.R
\name{scatter_theme}
\alias{scatter_theme}
\title{scatter theme}
\source{
modified from https://stackoverflow.com/questions/17753101/center-x-and-y-axis-with-ggplot2
}
\usage{
scatter_theme(p, axis_begin, axis_end, total_ticks = 11)
}
\arguments{
\item{p}{the ggplot to be converted}

\item{axis_begin}{the x and y limits start of the axis}

\item{axis_end}{the x and y limits end of the axis.}

\item{total_ticks}{total ticks on each axis. Ideally an uneven number.}
}
\description{
Setups a centered x-y axis scatter plot with label ticks. Note,
that No aes will be inherited from the original ggplot.
}
\note{
gridlines can be entirely removed by subsequently calling \code{theme_void()}.
}
\seealso{
Other ggplot: 
\code{\link{plot_tissue_enrichment}()}
}
\concept{ggplot}
