% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_names_by_dataset.R
\name{set_names_by_dataset}
\alias{set_names_by_dataset}
\title{set_names_by_dataset}
\usage{
set_names_by_dataset(data, overlay, marker = "color", by = "dataset")
}
\arguments{
\item{data}{a data.frame object.}

\item{overlay}{a data.frame object.}

\item{marker}{what should be organized?}

\item{by}{what should marker be organized by?}
}
\description{
the marker (e.g. colors) supplied to plotly must be in a global 
scheme with appropiate names (similar to a dict). This function generates
the appropiate mapping, so that the color scheme can be correctly plotted.
}
\seealso{
Other interactivity: 
\code{\link{make_interactive}()}
}
\concept{interactivity}
