% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.R
\name{standard_reference}
\alias{standard_reference}
\title{format to standard reference}
\usage{
standard_reference(ref)
}
\arguments{
\item{ref}{a reference.data.frame or list of data.frames}
}
\description{
formats a reference data.frame to contain
all needed collumns so that they can be merged in the future.
}
