% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{strsplit.nchar}
\alias{strsplit.nchar}
\title{strsplit.nchar}
\usage{
strsplit.nchar(x, nchar, suffix = "...")
}
\arguments{
\item{x}{a vector of strings.}

\item{nchar}{integer.}

\item{suffix}{string. what should all strings end with?}
}
\description{
split a string by a charcater
}
