% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_snp_loci.R
\name{subset_snp_loci}
\alias{subset_snp_loci}
\title{Subset SNP-to-gene list}
\usage{
subset_snp_loci(df)
}
\arguments{
\item{df}{data.frame containing SNP and gene columns}
}
\value{
list of two data.frames (singleGeneDf or multiGeneDf) corresponding to single-gene or multi-gene loci. Each contains SNP and gene columns.
}
\description{
Subset a SNP-to-gene list into single-gene and multi-gene loci
}
