% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{symbol_to_shape}
\alias{symbol_to_shape}
\title{translate plotly symbols to ggplot shapes}
\usage{
symbol_to_shape(vec)
}
\arguments{
\item{vec}{a character vector of plotly symbols.}
}
\description{
translate plotly symbols to ggplot shapes
}
\seealso{
Other misc: 
\code{\link{\%nin\%}()},
\code{\link{as.bait}()},
\code{\link{bold}()},
\code{\link{catf}()},
\code{\link{color_distinct}()},
\code{\link{color_gradient}()},
\code{\link{enumerate_replicate_combinations}()},
\code{\link{italics}()},
\code{\link{lun}()},
\code{\link{null_omit}()},
\code{\link{shape_to_symbol}()}
}
\concept{misc}
