% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{table_symbols}
\alias{table_symbols}
\title{get table of plotly/ggplot symbols}
\usage{
table_symbols()
}
\description{
get a table for translating
between plotly symbols and ggplot shapes.
}
\concept{misc
circle', # 13
square', # 15
circle', # 16
triangle', # 17
diamond', # 18
circle', # 19
circle', # 20}
