% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_markers.R
\name{tabulate_markers}
\alias{tabulate_markers}
\title{tabulate markers}
\usage{
tabulate_markers(lst, marker, by)
}
\arguments{
\item{lst}{a data.frame or a list of data.frames that all contain the arguments
'marker' and 'by' as column names.}

\item{marker}{what should be organized?}

\item{by}{what should marker be organized by?}
}
\value{
a table
}
\description{
a function that tabulates a data.frame or list of data.frames
by a certain marker. This means, that it will find all unique mappings from 'marker'
to 'by' in the form of a table.
}
\seealso{
Other interactivity: 
\code{\link{make_interactive}()},
\code{\link{set_names_by_dataset}()}
}
\concept{interactivity}
