% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_genoppi.R
\name{theme_genoppi}
\alias{theme_genoppi}
\title{Genoppi GG theme}
\usage{
theme_genoppi()
}
\description{
Set's up the classic genoppi theme for ggplot2.
}
\seealso{
Other ggplot: 
\code{\link{plot_tissue_enrichment}()},
\code{\link{scatter_theme}()},
\code{\link{theme_genoppi_bar}()},
\code{\link{volcano_theme}()}
}
\concept{ggplot}
