% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_genoppi.R
\name{theme_genoppi_bar}
\alias{theme_genoppi_bar}
\title{genoppi themed bar}
\usage{
theme_genoppi_bar(rotate = F)
}
\arguments{
\item{rotate}{boolean. Rotates axis titles.}
}
\description{
genoppi themed bar for shiny
}
