% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_genoppi.R
\name{theme_scatter}
\alias{theme_scatter}
\title{volcano theme}
\usage{
theme_scatter(p, axes_width = 0.22, axes_alpha = 0.5)
}
\arguments{
\item{p}{the ggplot to be converted}

\item{axes_width}{the width of the x and y-axis.}

\item{axes_alpha}{numeric between 0 and 1.}
}
\description{
Sets up a simple scatter theme by drawing x = 0 and y = 0 lines.
}
\seealso{
Other ggplot: 
\code{\link{plot_tissue_enrichment}()},
\code{\link{theme_genoppi_bar}()},
\code{\link{theme_volcano_custom}()},
\code{\link{theme_volcano}()}
}
\concept{ggplot}
