% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_genoppi.R
\name{theme_volcano}
\alias{theme_volcano}
\title{Volcano plot simple theme}
\usage{
theme_volcano(p)
}
\arguments{
\item{p}{the ggplot to be converted}
}
\description{
Set's up the classic genoppi theme for ggplot2.
}
\seealso{
Other ggplot: 
\code{\link{plot_tissue_enrichment}()},
\code{\link{theme_genoppi_bar}()},
\code{\link{theme_scatter}()},
\code{\link{theme_volcano_custom}()}
}
\concept{ggplot}
