% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_genoppi.R
\name{theme_volcano_custom}
\alias{theme_volcano_custom}
\title{volcano theme complex}
\usage{
theme_volcano_custom(
  p,
  tick_sz_x = -0.15,
  tick_sz_y = NULL,
  normalize_tick_sz = T,
  axes_width = 0.5,
  ticks_width = 0.7,
  ticks_labs_cex = 4,
  ticks_labs_x_vjust = 2,
  ticks_labs_y_hjust = 2,
  lab_cex = 2,
  xlab_vjust = 0.6,
  xlab_hjust = 0.38,
  ylab_vjust = 1.5,
  ylab_hjust = 0,
  font_family = "Helvetica",
  grid_lines = F,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{p}{the ggplot to be converted}

\item{tick_sz_x}{Tick-size.}

\item{tick_sz_y}{tick size.}

\item{normalize_tick_sz}{Boolean. If TRUE, the tick_sz_y will be normalized according to argument tick_sz_x.}

\item{axes_width}{the width of the x and y-axis.}

\item{ticks_width}{the width of the ticks.}

\item{ticks_labs_cex}{the character expansion factor of tick labels.}

\item{ticks_labs_x_vjust}{Vertical adjustment of tick labels on x axis.}

\item{ticks_labs_y_hjust}{Horizontal adjustment of tick labels on y axis.}

\item{lab_cex}{numeric. Character expansion factor for labels.}

\item{xlab_vjust}{Vertical adjustment on label on x-axis.}

\item{xlab_hjust}{Horizontal adjustment on label on x-axis.}

\item{ylab_vjust}{Vertical adjustment on label on y-axis.}

\item{ylab_hjust}{Horizontal adjustment on label on y-axis.}

\item{font_family}{string for the font family.}

\item{grid_lines}{Boolean. Draw grid lines?}

\item{xlab}{String. alternative x label.}

\item{ylab}{String. alternative y label.}
}
\description{
Sets up custom volcano plot theme.
}
\seealso{
Other ggplot: 
\code{\link{plot_tissue_enrichment}()},
\code{\link{theme_genoppi_bar}()},
\code{\link{theme_scatter}()},
\code{\link{theme_volcano}()}
}
\concept{ggplot}
