% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_overlay_data.R
\name{to_overlay_data}
\alias{to_overlay_data}
\title{to overlay data}
\usage{
to_overlay_data(df, dataset = NULL)
}
\arguments{
\item{df}{a data.frame}

\item{dataset}{optional string, if dataset is not provided in df.}
}
\description{
Adds columns to a data.frame (e.g. one used for overlay)
that are required for downstream functionality in genoppi/
}
\concept{interactive}
