% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_reference.R
\name{validate_reference}
\alias{validate_reference}
\title{validate reference data.frame}
\usage{
validate_reference(
  df,
  valid = c("gene", "accession_number", "col_significant", "col_other", "shape",
    "dataset", "stroke", "alt_label", "label", "label_size", "size", "symbol", "pLI",
    "opacity", "size", "gg.size", "legend_order")
)
}
\arguments{
\item{df}{a data.frame}

\item{valid}{a vector of valid downstream ggplot/plotly.}
}
\value{
a data.frame
}
\description{
Input a data.frame and check whether it contains a pre-specified
set of valid columns. Invalid columns will be omitted alongside a warning.
}
\concept{overlay}
