% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn_to_table.R
\name{venn_to_table}
\alias{venn_to_table}
\title{venn diagram to table}
\usage{
venn_to_table(venn)
}
\arguments{
\item{venn}{a list of length two that each contains a vector of genes}

\item{datnames}{a vector of string with the names of the venn items}
}
\description{
converts a venn diagram (generated from draw_genoppi_venn)
into a table that can be exported.
}
