% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{vline}
\alias{vline}
\title{vertical line}
\usage{
vline(x = 0, color = "red", width = 1, dash = "dash")
}
\arguments{
\item{x}{x-value, numeric.}

\item{color}{color, string.}

\item{width}{line width, numeric.}

\item{dash}{string. 'solid, 'dot' or 'dash'.}
}
\description{
vertical plotly line
}
\seealso{
Other misc: 
\code{\link{\%nin\%}()},
\code{\link{as.bait}()},
\code{\link{bold}()},
\code{\link{catf}()},
\code{\link{color_distinct}()},
\code{\link{color_gradient}()},
\code{\link{enumerate_replicate_combinations}()},
\code{\link{get_gg_legend}()},
\code{\link{hline}()},
\code{\link{italics}()},
\code{\link{lun}()},
\code{\link{null_omit}()},
\code{\link{shape_to_symbol}()},
\code{\link{symbol_to_shape}()}
}
\concept{misc}
