% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_genoppi.R
\name{volcano_theme}
\alias{volcano_theme}
\title{volcano theme}
\source{
modified from https://stackoverflow.com/questions/17753101/center-x-and-y-axis-with-ggplot2
}
\usage{
volcano_theme(
  p,
  xlims = c(-10, 10),
  ylims = c(0, 5),
  x_ticks = 11,
  y_ticks = 11,
  grid_width = 1
)
}
\arguments{
\item{p}{the ggplot to be converted}

\item{xlims}{vector. the lower and upper bound of x-axis.}

\item{ylims}{vector. the lower and upper bound of y-axis.}

\item{x_ticks}{total ticks on x-axis.}

\item{y_ticks}{total ticks on y-axis.}

\item{grid_width}{the width of the grid.}
}
\description{
Setups a volcano theme, with a yaxis at x=0, and a grid.
}
\note{
gridlines can be entirely removed by subsequently calling \code{theme_void()}.
}
\seealso{
Other ggplot: 
\code{\link{plot_tissue_enrichment}()},
\code{\link{scatter_theme}()},
\code{\link{theme_genoppi_bar}()},
\code{\link{theme_genoppi}()}
}
\concept{ggplot}
