% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accession_gene_table}
\alias{accession_gene_table}
\title{Uniprot protein accession number to HGNC gene symbol mapping}
\format{
A data frame with 67,519 rows and 2 variables:
\describe{
  \item{accession_number}{Uniprot accession number (e.g. P10415)}
  \item{gene}{HGNC gene symbol (e.g. BCL2)}
}
}
\source{
ID mapping data for human protein-coding genes downloaded from Uniprot:  
\url{https://www.uniprot.org}
}
\usage{
accession_gene_table
}
\description{
A dataset containing 67,519 Uniprot accession number to HGNC gene symbol pairings.
}
\references{
UniProt Consortium. UniProt: a worldwide hub of protein knowledge. Nucleic Acids Res.
2019;47(D1):D506-D515. doi:10.1093/nar/gky1049
}
\seealso{
Other datasets: 
\code{\link{bioplex_table}},
\code{\link{example_data2}},
\code{\link{example_data3}},
\code{\link{example_data}},
\code{\link{genes_snps}},
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gtex_protein}},
\code{\link{gtex_rna}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{hpa_rna}},
\code{\link{inweb_table}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c2_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_h_table}}
}
\concept{datasets}
\keyword{datasets}
