% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accession_gene_table}
\alias{accession_gene_table}
\title{gene ID mapping table}
\format{data.frame object with columns: accession_number, gene}
\source{
Protein-coding genes downloaded from Uniprot.
}
\usage{
accession_gene_table
}
\description{
A table containing 67,744 entries of accession_number to gene (HGNC symbol) mapping.
}
\seealso{
Other datasets: 
\code{\link{example_data}},
\code{\link{genes_snps}},
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{inweb_hash}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c2_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_h_table}}
}
\concept{datasets}
\keyword{datasets}
