% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_genoppi_trace.R
\name{add_genoppi_trace}
\alias{add_genoppi_trace}
\title{add genoppi trace}
\usage{
add_genoppi_trace(
  p,
  data,
  parameters,
  stroke_width = 0.2,
  legend = F,
  legend_group = ""
)
}
\arguments{
\item{p}{a plot_ly() object.}

\item{data}{a data.frame with columns dataset, gene,}

\item{parameters}{a list of parameters or an environement generated with \code{environment()}.}
}
\description{
adding a trace with new data to plotly is non-trivial. This function takes
care of adding data correctly with the correct coloring.
}
\note{
global variable 'global_colors' must be specified using setNames().
}
