% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hover_lines_volcano.R
\name{add_hover_lines_volcano}
\alias{add_hover_lines_volcano}
\title{add hover lines}
\usage{
add_hover_lines_volcano(
  p,
  line_pvalue,
  line_logfc,
  logfc_direction = "both",
  sig_type = "fdr"
)
}
\arguments{
\item{p}{a ggplot}

\item{line_pvalue}{the pvalue threshold}

\item{line_logfc}{the logfc threshold}

\item{logfc_direction}{string. Can be either \code{'positive'}, \code{'negative'} or \code{'both'}.}

\item{sig_type}{string. Can be either \code{'fdr'} or \code{'pvalue'}.}
}
\description{
A function that adds reactive dashed lines to a plotly object.
}
\seealso{
Other shiny: 
\code{\link{add_layout_html_axes_scatterplot}()},
\code{\link{add_layout_html_axes_volcano}()},
\code{\link{add_markers_search_pathway}()},
\code{\link{add_markers_search}()}
}
\concept{shiny}
