% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_layout_html_axes.R
\name{add_layout_html_axes_scatterplot}
\alias{add_layout_html_axes_scatterplot}
\title{layout for scatterplot}
\usage{
add_layout_html_axes_scatterplot(
  p,
  repA,
  repB,
  title = "",
  legend_title = bold("Overlay")
)
}
\arguments{
\item{p}{a ggplot}
}
\description{
format axes.
}
\seealso{
Other shiny: 
\code{\link{add_hover_lines_volcano}()},
\code{\link{add_layout_html_axes_volcano}()},
\code{\link{add_markers_search_pathway}()},
\code{\link{add_markers_search}()}
}
\concept{shiny}
