% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_layout_html_axes.R
\name{add_layout_html_axes_volcano}
\alias{add_layout_html_axes_volcano}
\title{layout for volcano plot.}
\usage{
add_layout_html_axes_volcano(p, height = 400, width = 400)
}
\arguments{
\item{p}{a ggplot}
}
\description{
Format axes.
}
\note{
internal
}
\seealso{
Other shiny: 
\code{\link{add_hover_lines_volcano}()},
\code{\link{add_layout_html_axes_scatterplot}()},
\code{\link{add_markers_search}()}
}
\concept{shiny}
