% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_layout_html_axes.R
\name{add_layout_html_axes_volcano}
\alias{add_layout_html_axes_volcano}
\title{layout for volcano plot.}
\usage{
add_layout_html_axes_volcano(
  p,
  height = NULL,
  width = NULL,
  legend_title = bold("Overlay")
)
}
\arguments{
\item{p}{a plotly object.}

\item{height}{the heigh of the plotly object.}

\item{width}{the width of the plotly object.}

\item{legend_title}{the title of the legend.}
}
\description{
Format axes.
}
\seealso{
Other shiny: 
\code{\link{add_hover_lines_volcano}()},
\code{\link{add_layout_html_axes_scatterplot}()},
\code{\link{add_markers_search_pathway}()},
\code{\link{add_markers_search}()}
}
\concept{shiny}
