% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_markers_search.R
\name{add_markers_search}
\alias{add_markers_search}
\title{search volcano plots}
\usage{
add_markers_search(p, genes)
}
\arguments{
\item{p}{a ggplot}

\item{genes}{a vector of genes}
}
\description{
Adds search markers to a plot. Requires a plotly object.
}
\note{
internal
}
\seealso{
Other shiny: 
\code{\link{add_hover_lines_volcano}()},
\code{\link{add_layout_html_axes_scatterplot}()},
\code{\link{add_layout_html_axes_volcano}()},
\code{\link{add_markers_search_pathway}()}
}
\concept{shiny}
