% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_to_column.R
\name{append_to_column}
\alias{append_to_column}
\title{Append a string to a column}
\usage{
append_to_column(data, sig_text = "(enriched)", insig_text = "(not enriched)")
}
\arguments{
\item{data}{a data or overlay data.frame with \code{dataset} as column.}

\item{sig_text}{what should be the significance text?}

\item{insig_text}{what should be the insignificance text?}
}
\description{
Appends a dynamic string to a column condition on whether
the another column 'significant' is true or false.
}
\note{
In order to correctly draw plotly points
the identifier (dataset) must have a unique colors associated with it.
This function appends the dataset column with a significance text.
}
