% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{as.bait}
\alias{as.bait}
\title{as.bait}
\usage{
as.bait(bait)
}
\arguments{
\item{bait}{string indicating the bait.}
}
\description{
quickly format the bait so that it can be used by various overlay functions.
}
