% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{as.bait}
\alias{as.bait}
\title{as.bait}
\usage{
as.bait(bait)
}
\arguments{
\item{bait}{string indicating the bait.}
}
\description{
quickly format the bait so that it can be used by various overlay functions.
}
\seealso{
Other misc: 
\code{\link{\%nin\%}()},
\code{\link{bold}()},
\code{\link{catf}()},
\code{\link{color_distinct}()},
\code{\link{color_gradient}()},
\code{\link{enumerate_replicate_combinations}()},
\code{\link{italics}()},
\code{\link{lun}()},
\code{\link{null_omit}()}
}
\concept{misc}
