% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{assign_color}
\alias{assign_color}
\title{assign color by column}
\usage{
assign_color(df, col, by_freq = T, palette = color_distinct())
}
\arguments{
\item{df}{a data.frame}

\item{col}{the identifiying column for assiging color}
}
\description{
assigns color to a data.frame by a certain column
}
