% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bioplex_table}
\alias{bioplex_table}
\title{Bioplex 3.0 protein-protein interactions}
\format{
A data.frame with 883,356 rows and 4 variables:
\describe{
  \item{Gene1}{HGNC gene symbol of first protein (bait)}
  \item{Gene2}{HGNC gene symbol of second protein (prey)}
  \item{pW}{The probability that the bait-prey pair corresponds to a wrong identification.}
  \item{pNI}{The probability of the bait-prey pair being a background protein.}
  \item{pInt}{The probability of the bait-prey being high Confidence interacting proteins.}
}
}
\source{
See cited references; downloaded 4-Jan-2021 from \url{https://bioplex.hms.harvard.edu/interactions.php}.
}
\usage{
bioplex_table
}
\description{
A dataset containing 118,162 protein-protein interactions. Generated through AP-MS of 
10,128 baits in HEK293T cell lines. See the source for more details about included columns.
}
\references{
Huttlin et al. Dual Proteome-scale Networks Reveal Cell-specific Remodeling of the Human Interactome.
2020 bioRxiv doi: 10.1101/2020.01.19.905109.
}
\seealso{
Other datasets: 
\code{\link{accession_gene_table}},
\code{\link{example_data2}},
\code{\link{example_data3}},
\code{\link{example_data}},
\code{\link{genes_snps}},
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gtex_protein}},
\code{\link{gtex_rna}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{hpa_rna}},
\code{\link{inweb_table}},
\code{\link{irefindex_table}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c2_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_h_table}}
}
\concept{datasets}
\keyword{datasets}
