% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bioplex_table}
\alias{bioplex_table}
\title{BioPlex protein-protein interactions}
\format{
A data.frame with 883,356 rows and 4 variables:
\describe{
  \item{Gene1}{HGNC gene symbol of first protein (bait)}
  \item{Gene2}{HGNC gene symbol of second protein (prey)}
  \item{pW}{Probability that the prey is a wrong identification}
  \item{pNI}{Probability of the prey is a background protein}
  \item{pInt}{Probability of the prey is a high-confidence interacting protein}
}
}
\source{
BioPlex data (v3.0) downloaded on 2021-01-04 from:
\url{https://bioplex.hms.harvard.edu/interactions.php}
}
\usage{
bioplex_table
}
\description{
A dataset containing 118,162 protein-protein interactions in BioPlex 3.0,
identified through AP-MS of 10,128 human bait proteins in HEK293T cells.
}
\references{
Huttlin EL, Ting L, Bruckner RJ, et al. The BioPlex Network: A Systematic Exploration of the Human Interactome.
Cell. 2015;162(2):425-440. doi:10.1016/j.cell.2015.06.043

Huttlin EL, Bruckner RJ, Navarrete-Perea J, et al. Dual Proteome-scale Networks Reveal Cell-specific Remodeling
of the Human Interactome. bioRxiv. 2020. doi:10.1101/2020.01.19.905109
}
\seealso{
Other datasets: 
\code{\link{accession_gene_table}},
\code{\link{example_data2}},
\code{\link{example_data3}},
\code{\link{example_data}},
\code{\link{genes_snps}},
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gtex_brain_table}},
\code{\link{gtex_protein_table}},
\code{\link{gtex_rna_table}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{hpa_rna_table}},
\code{\link{huri_table}},
\code{\link{inweb_table}},
\code{\link{irefindex_table}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c2_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_c8_table}},
\code{\link{msigdb_h_table}},
\code{\link{pcnet_table}},
\code{\link{string_table}}
}
\concept{datasets}
\keyword{datasets}
