% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{bold}
\alias{bold}
\title{bold}
\usage{
bold(x)
}
\arguments{
\item{x}{string}
}
\description{
make text html bold
}
\seealso{
Other misc: 
\code{\link{\%nin\%}()},
\code{\link{as.bait}()},
\code{\link{catf}()},
\code{\link{color_distinct}()},
\code{\link{color_gradient}()},
\code{\link{enumerate_replicate_combinations}()},
\code{\link{italics}()},
\code{\link{lun}()},
\code{\link{null_omit}()}
}
\concept{misc}
