% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_adujusted_enrichment.R
\name{calc_adjusted_enrichment}
\alias{calc_adjusted_enrichment}
\title{calculate adjusted enrichment}
\usage{
calc_adjusted_enrichment(data, reference, col.by = "tissue", bait = NULL)
}
\arguments{
\item{data}{proteomic data with gene and significant columns.}

\item{reference}{table with col.by, gene, significant in columns.}

\item{col.by}{string. What column contains the group?}

\item{bait}{bait. String. Passed to \code{?calc_hyper}.}
}
\description{
calculate benjamin-hochberg adjusted enrichment of genes in tissue.
}
\examples{
\dontrun{
# check for enrichment in GTEx
data(GTEX_table)
data(example_data)
data = example_data \%>\% calc_mod_ttest \%>\% id_enriched_proteins()
gtex_enrichment = calc_adjusted_enrichment(data, GTEX_table)
gtex_enrichment
}
}
