% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cumsum_table.R
\name{calc_cumsum_table}
\alias{calc_cumsum_table}
\title{calculate the cumulative sum of frequencies}
\usage{
calc_cumsum_table(x, col.id, col.freq = "Freq")
}
\arguments{
\item{x}{a data.frame containign col.id and col.freq}

\item{col.id}{string. what item in \code{x} is the identifier?}

\item{col.freq}{string. what numeric column contains the frequency?}
}
\description{
calculate the cumulative sum of frequencies
}
\keyword{internal}
