% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hyper.R
\name{calc_hyper}
\alias{calc_hyper}
\title{Perform hypergeometric test}
\usage{
calc_hyper(df, listDf, intersectDf, bait = NULL)
}
\arguments{
\item{df}{data.frame containing proteomic data, with gene and significant columns}

\item{listDf}{data.frame containing data to overlap, with listName, gene and significant columns}

\item{intersectDf}{data.frame contaning listName and intersectN columns intersectN = vector of boolean variables indicating if total population should be intersect of the two datasets}

\item{bait}{name of bait protein}
}
\value{
list of data.frame and list. Data.frame contains list_name, successInSample_count (x), sample_count (n), notSample_count (N-n), success_count (k), population_count (N), pvalue. List of lists (for each listName) contains genes names corresponding to each group (successInSample_genes, sample_genes, notSample_genes, success_genes, population_genes).
}
\description{
Use one-tailed hypergeometric test to calcualte overlap enrichment between proteomic data and other datasets (e.g. InWeb, gene lists, etc.)
}
\examples{
\dontrun{
# an example

}
}
