% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mod_ttest.R
\name{calc_mod_ttest}
\alias{calc_mod_ttest}
\title{Perform one-sample moderated t-test}
\usage{
calc_mod_ttest(df, iter = 2000, order = T)
}
\arguments{
\item{df}{data.frame containing gene and rep[0-9] (replicate logFC) columns.}

\item{iter}{integer indicating maximum number of iterations to perform in limma::limFit().}

\item{order}{boolean. Should the result be ordered by logFC and FDR?}
}
\value{
data.frame containing input df + logFC, pvalue, and FDR columns; sorted by decreasing logFC, then FDR.
}
\description{
Use one-sample moderated t-test implemented in limma package to calculate protein logFC, pvalue, and FDR.
See \code{?limma::lmFit} and \code{?limma::eBayes} for more details.
}
\examples{
\dontrun{
data("example_data")
stats_df <- calc_mod_ttest(example_data) 
}

}
