% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mod_ttest.R
\name{calc_mod_ttest}
\alias{calc_mod_ttest}
\title{Perform moderated t-test}
\usage{
calc_mod_ttest(df)
}
\arguments{
\item{df}{data.frame containing gene and rep (replicate logFC)  columns}
}
\value{
data.frame containing containing df + logFC, pvalue, and FDR columns; sorted by decreasing logFC, then FDR
}
\description{
Use moderated t-test implemented in limma package to calculate logFC, pvalue, and FDR.
See \code{?limma::lmFit} and \code{?limma::eBayes} for more details.
}
\examples{
\dontrun{
data("example_data")
stats_df <- calc_mod_ttest(example_data) 
}

}
