% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{catf}
\alias{catf}
\title{warnings to stderr}
\usage{
catf(msg, file = stderr())
}
\arguments{
\item{msg}{the message}

\item{file}{string, e.g. a filename.}
}
\description{
sends a message to stderr (i.e shiny)
}
\seealso{
Other misc: 
\code{\link{\%nin\%}()},
\code{\link{as.bait}()},
\code{\link{bold}()},
\code{\link{color_distinct}()},
\code{\link{color_gradient}()},
\code{\link{enumerate_replicate_combinations}()},
\code{\link{italics}()},
\code{\link{lun}()},
\code{\link{null_omit}()}
}
\concept{misc}
