% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_input}
\alias{check_input}
\title{Check input data format}
\usage{
check_input(tabl)
}
\arguments{
\item{tabl}{a data.frame}
}
\value{
list of boolean vairables indicating if input match a format and allowed formats
}
\description{
Check if input data columns match one of the four allowed formats.
}
\note{
An item in the approved format list can appear more than once
}
\examples{
\dontrun{
tabl1 = data.frame(gene = letters,rep1=runif(26), rep2=rnorm(26))
tabl2 = data.frame(gene = letters,rep1=runif(26), rep2=rnorm(26), somethingelse=1:26)
check_input(tabl1)
}

}
