% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_labels.R
\name{collapse_labels}
\alias{collapse_labels}
\title{collaps overlay data}
\usage{
collapse_labels(overlay)
}
\arguments{
\item{overlay}{a data.frame}
}
\description{
a data.frame used to overlay items in genoppi
will look for non-unqiue gene names and combine them into 
a single informative line using their alt_label (e.g. snp)
and their corrsponding dataset.
}
