% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{color_distinct}
\alias{color_distinct}
\title{distinct coloring}
\usage{
color_distinct()
}
\description{
generates vector of 74 distinct colors from RColorBrewer.
}
\seealso{
Other misc: 
\code{\link{\%nin\%}()},
\code{\link{as.bait}()},
\code{\link{bold}()},
\code{\link{catf}()},
\code{\link{color_gradient}()},
\code{\link{enumerate_replicate_combinations}()},
\code{\link{italics}()},
\code{\link{lun}()},
\code{\link{null_omit}()}
}
\concept{misc}
