% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{color_gradient}
\alias{color_gradient}
\title{color gradient}
\source{
stackoverflow snippet
}
\usage{
color_gradient(x, colors = c("green", "red"), colsteps = 100)
}
\arguments{
\item{x}{vector of values}
}
\description{
makes a function for getting gradient colors
}
