% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{color_gradient}
\alias{color_gradient}
\title{color gradient}
\source{
stackoverflow snippet
}
\usage{
color_gradient(x, colors = c("green", "red"), colsteps = 100)
}
\arguments{
\item{x}{vector of values}

\item{colors}{colors from which to form gradient.}

\item{colsteps}{how many colors should be created?}
}
\description{
makes a function for getting gradient colors
}
\seealso{
Other misc: 
\code{\link{\%nin\%}()},
\code{\link{as.bait}()},
\code{\link{bold}()},
\code{\link{catf}()},
\code{\link{color_distinct}()},
\code{\link{enumerate_replicate_combinations}()},
\code{\link{italics}()},
\code{\link{lun}()},
\code{\link{null_omit}()}
}
\concept{misc}
