% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_genoppi_venn.R
\name{draw_genoppi_venn}
\alias{draw_genoppi_venn}
\title{Draw venn diagrams}
\usage{
draw_genoppi_venn(x, main = "", colors = c("cornflowerblue", "yellow1"))
}
\arguments{
\item{x}{a list of two vectors that contain some overlapping strings.}

\item{main}{title, typically a p-value.}

\item{colors}{color scheme.}
}
\description{
drwas a simple 1 by 1 overlap venn diagram.
}
\note{
NAs in x will be discarded and produce a warning.
}
