% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter_basic.R
\name{enumerate_replicate_combinations}
\alias{enumerate_replicate_combinations}
\title{enumerate replicate combinations}
\usage{
enumerate_replicate_combinations(len)
}
\arguments{
\item{len}{number of replicates}
}
\description{
enumerates unique replicate comparisons
}
\concept{misc}
