% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate_replicate_combinations.R
\name{enumerate_replicate_combinations}
\alias{enumerate_replicate_combinations}
\title{enumerate replicate combinations}
\usage{
enumerate_replicate_combinations(len)
}
\arguments{
\item{len}{number of replicates}
}
\description{
enumerates unique replicate comparisons
}
\seealso{
Other misc: 
\code{\link{\%nin\%}()},
\code{\link{as.bait}()},
\code{\link{bold}()},
\code{\link{catf}()},
\code{\link{color_distinct}()},
\code{\link{color_gradient}()},
\code{\link{italics}()},
\code{\link{lun}()},
\code{\link{null_omit}()}
}
\concept{misc}
