% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example proteomic data (1)}
\format{
A data frame with 556 rows and 4 variables:
\describe{
  \item{gene}{HGNC gene symbol}
  \item{rep1}{replicate 1 log2 fold change}
  \item{rep2}{replicate 2 log2 fold change}
  \item{rep3}{replicate 3 log2 fold change}
}
}
\source{
Genoppi IP-MS experiment
}
\usage{
example_data
}
\description{
A dataset containing gene symbol and corresponding log2 fold change values for 3 replicates,
derived from BCL2 (bait) vs. IgG (control) immunoprecipitation-mass spectrometry experiment in a neuron cell line (GPiN).
}
\references{
Pintacuda G, Lassen FH, Hsu YH, Kim A, Martín JM, Malolepsza E et al. Genoppi is an open-source software
for robust and standardized integration of proteomic and genetic data. Nat Commun. 2021;12(1):2580.
doi:10.1038/s41467-021-22648-5
}
\seealso{
Other datasets: 
\code{\link{accession_gene_table}},
\code{\link{bioplex_table}},
\code{\link{example_data2}},
\code{\link{example_data3}},
\code{\link{genes_snps}},
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gtex_protein}},
\code{\link{gtex_rna}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{hpa_rna}},
\code{\link{inweb_table}},
\code{\link{irefindex_table}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c2_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_h_table}}
}
\concept{datasets}
\keyword{datasets}
