% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_snps_data.R
\docType{data}
\name{genes_snps}
\alias{genes_snps}
\title{SNP-to-gene mapping hash table}
\format{hash object with gene name as key, vector of SNPs as value}
\source{
"Genoppi-v3 snp_to_gene.Rdata file created by April Kim"
}
\usage{
genes_snps
}
\description{
A hash table containing 18,116 genes mapped to LD loci of 1KG phase 3 SNPs.
}
\keyword{datasets}
