% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes_snps}
\alias{genes_snps}
\title{SNP-to-gene mapping hash table}
\format{hash object with gene name as key, vector of SNPs as value}
\source{
Genoppi-v3 snp_to_gene.Rdata file created by April Kim.
}
\usage{
genes_snps
}
\description{
A hash table containing 18,116 genes mapped to LD loci of 1KG phase 3 SNPs.
}
\seealso{
Other datasets: 
\code{\link{accession_gene_table}},
\code{\link{example_data}},
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{inweb_hash}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c2_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_h_table}}
}
\concept{datasets}
\keyword{datasets}
