% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes_snps_hash}
\alias{genes_snps_hash}
\title{SNP-to-gene mapping}
\format{
A hash object with 19,326 gene names as keys and vectors of SNP IDs as values
}
\source{
GENCODE v43 data downloaded from:
\url{https://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_43/gencode.v43.basic.annotation.gtf.gz}
1000 Genomes phase 3 data (with dbSNP build 156 annotation) downloaded from:
\url{https://www.cog-genomics.org/plink/2.0/resources#phase3_1kg}
}
\usage{
genes_snps_hash
}
\description{
A dataset containing 19,326 protein-coding genes (GENCODE v43) mapped to SNP IDs (dbSNP build 156).
A gene is linked to a SNP if its genomic coordinates overlap with linkage disequilibrium (LD) locus
boundaries of the SNP. LD loci were calculated using the 1000 Genomes phase 3 EUR panel and PLINK,
defined as regions spanning tagging SNPs (r2 > 0.6) ± 50kb; SNPs with MAF < 0.05 were excluded.
}
\references{
#
Frankish A, Diekhans M, Jungreis I, et al. GENCODE 2021.
Nucleic Acids Res. 2021;49(D1):D916-D923. doi:10.1093/nar/gkaa1087

Byrska-Bishop M, Evani US, Zhao X, et al. High-coverage whole-genome sequencing of the expanded 1000 Genomes
Project cohort including 602 trios. Cell. 2022;185(18):3426-3440.e19. doi:10.1016/j.cell.2022.08.004

Chang CC, Chow CC, Tellier LC, Vattikuti S, Purcell SM, Lee JJ. Second-generation PLINK: rising to the challenge
of larger and richer datasets. Gigascience. 2015;4:7. Published 2015 Feb 25. doi:10.1186/s13742-015-0047-8
}
\seealso{
Other datasets: 
\code{\link{accession_gene_table}},
\code{\link{bioplex_table}},
\code{\link{example_data2}},
\code{\link{example_data3}},
\code{\link{example_data}},
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gtex_brain_table}},
\code{\link{gtex_protein_table}},
\code{\link{gtex_rna_table}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{hpa_rna_table}},
\code{\link{inweb_table}},
\code{\link{irefindex_table}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c2_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_c8_table}},
\code{\link{msigdb_h_table}},
\code{\link{string_table}}
}
\concept{datasets}
\keyword{datasets}
