% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bioplex_list.R
\name{get_bioplex_list}
\alias{get_bioplex_list}
\title{Retrieve Bioplex 3.0 interactors for a given bait}
\usage{
get_bioplex_list(bait, p = 0.9)
}
\arguments{
\item{bait}{string. name of bait protein}

\item{p}{numeric. Probability of the protein being an interactor with the bait. See \code{?bioplex_table}.}
}
\value{
data.frame containing gene and significant columns for all non-bait bioplex genes 
(significant=T for bioplex 3.0 interactors of bait). NULL if bait not found in bioplex.
}
\description{
Use bioplex_table data to get bioplex interactors and non-interactors of bait. 
See \code{?bioplex_table} for more details about the data set.
}
\examples{
\dontrun{
df1 <- get_bioplex_list('BCL2',p = 0.5)
}
}
