% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneset_overlay.R
\name{get_geneset_overlay}
\alias{get_geneset_overlay}
\title{get geneset as overlay}
\usage{
get_geneset_overlay(data, database, k = 100, only.significant = T)
}
\arguments{
\item{database}{see get_pathways}

\item{k}{integer. Up to how many of the most recurrent gene sets should be displayed at once?}

\item{only.significant}{boolean. If true, only column with significant=TRUE will be considered.}

\item{pulldown}{pulldown data}
}
\description{
generates a list of data.frames that can be inputted to \code{plot_overlay()}
}
\note{
this is a work in progress.
}
\examples{
\dontrun{
data("example_data")
df = example_data \%>\% 
   calc_mod_ttest() \%>\% 
   id_enriched_proteins()
   
plot_volcano_basic(df) \%>\% 
   plot_overlay(get_geneset_overlay(df, 'hgnc', k = 25)) \%>\% 
   make_interactive()
}
}
\keyword{internal}
