% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inweb_list.R
\name{get_inweb_list}
\alias{get_inweb_list}
\title{Retrieve InWeb data for a given bait}
\usage{
get_inweb_list(bait)
}
\arguments{
\item{bait}{name of bait protein}
}
\value{
data.frame containing gene and significant columns for all InWeb genes (significant=T for InWeb interactors of bait). NULL if bait not found in InWeb.
}
\description{
Use inweb_hash data to get InWeb interactors and non-interactors of bait.
}
