% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_irefindex_list.R
\name{get_irefindex_list}
\alias{get_irefindex_list}
\title{Retrieve irefindex for a given bait}
\usage{
get_irefindex_list(bait, n = 1)
}
\arguments{
\item{bait}{string. name of bait protein}

\item{n}{numeric. Minimum number of publications that this interaction has been described in.}
}
\value{
data.frame containing gene and significant columns for all non-bait IRefIndex genes 
(significant=T for IRefIndex interactors of bait). NULL if bait not found in IRefIndex.
}
\description{
Use irefindex_table data to get IRefIndex interactors and non-interactors of bait. 
See \code{?irefindex_table} for more details about the data set.
}
\examples{
\dontrun{
df1 <- get_irefindex_list('BCL2',n = 1)
}
}
