% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tissue_list.R
\name{get_tissue_list}
\alias{get_tissue_list}
\title{Retrieve tissue-specific expression data}
\usage{
get_tissue_list(tissue, table = gtex_rna_table)
}
\arguments{
\item{tissue}{vector of tissue names (character) included in dataset below.}

\item{table}{dataset name (One of \code{?gtex_rna_table}, \code{?gtex_brain_table}, 
\code{?gtex_protein_table}, \code{?hpa_rna_table}).}
}
\description{
Create gene list data.frame from input tissues that can be superimposed onto volcano plots
or used for statistical enrichment analysis.
}
\examples{
\dontrun{
example_data \%>\% 
  calc_mod_ttest() \%>\% 
  id_significant_proteins() \%>\% 
  plot_volcano_basic() \%>\% 
  plot_overlay(as.bait('BCL2')) \%>\%
  plot_overlay(list(brain = get_tissue_list("Brain - Cortex",table = gtex_rna))) \%>\%
  theme_volcano()
}
}
