% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tissue_lists.R
\name{get_tissue_lists}
\alias{get_tissue_lists}
\title{Retreive Human Protein Atlas specific data}
\usage{
get_tissue_lists(tissue, table = hpa_table)
}
\arguments{
\item{tissue}{vector of tissue names (character). See details for acceptable names.}

\item{table}{either hpa_table or gtex_table}
}
\description{
Create gene list data.frame from input traits
}
\details{
The following 33 tissues are available for Human Protein Atlas:
\itemize{
\item 'adipose tissue'
\item 'adrenal gland'
\item 'bone marrow'
\item 'breast'
\item 'cervix, uterine'
\item 'ductus deferens'
\item 'endometrium 1'
\item 'epididymis'
\item 'esophagus'
\item 'fallopian tube'
\item 'gallbladder'
\item 'heart muscle'
\item 'kidney'
\item 'liver'
\item 'lung'
\item 'ovary'
\item 'pancreas'
\item 'parathyroid gland'
\item 'pituitary gland'
\item 'placenta'
\item 'prostate'
\item 'retina'
\item 'salivary gland'
\item 'seminal vesicle'
\item 'skeletal muscle'
\item 'skin 1'
\item 'smooth muscle'
\item 'stomach 1'
\item 'testis'
\item 'thyroid gland'
\item 'tongue'
\item 'urinary bladder'
\item 'vagina'
}
}
