% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnomad_data.R
\docType{data}
\name{gnomad_table}
\alias{gnomad_table}
\title{gnomAD table}
\format{data.frame from gnomAD V 2.1.1}
\source{
https://storage.googleapis.com/gnomad-public/release/2.1.1/constraint/gnomad.v2.1.1.lof_metrics.by_gene.txt.bgz
}
\usage{
gnomad_table
}
\description{
The gnomAD v2 data set contains data from 125,748 exomes and 15,708 whole genomes, all mapped to the GRCh37/hg19 reference sequence.
}
\seealso{
Other datasets: 
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{msigdb_table}}
}
\concept{datasets}
\keyword{datasets}
