% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gnomad_table}
\alias{gnomad_table}
\title{gnomAD mutational constraint (pLI) annotations}
\format{
A data frame with 19,704 rows and 2 variables:
\describe{
  \item{gene}{HGNC gene symbol}
  \item{pLI}{pLI score}
}
}
\source{
pLoF Metrics by Gene dataset (v2.1.1) downloaded from:
\url{https://gnomad.broadinstitute.org/downloads#v2-constraint}
}
\usage{
gnomad_table
}
\description{
A dataset containing gnomAD pLI scores for 19,704 genes. The pLI score estimates the probability
that a gene is intolerant of loss-of-function mutations.
}
\references{
Lek M, Karczewski KJ, Minikel EV, et al. Analysis of protein-coding genetic variation in 60,706 humans.
Nature. 2016;536(7616):285-291. doi:10.1038/nature19057  

Karczewski KJ, Francioli LC, Tiao G, et al. The mutational constraint spectrum quantified from variation in 141,456 humans.
Nature. 2020;581(7809):434-443. doi:10.1038/s41586-020-2308-7
}
\seealso{
Other datasets: 
\code{\link{accession_gene_table}},
\code{\link{bioplex_table}},
\code{\link{example_data2}},
\code{\link{example_data3}},
\code{\link{example_data}},
\code{\link{genes_snps_hash}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gtex_brain_table}},
\code{\link{gtex_protein_table}},
\code{\link{gtex_rna_table}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{hpa_rna_table}},
\code{\link{huri_table}},
\code{\link{inweb_table}},
\code{\link{irefindex_table}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c2_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_c8_table}},
\code{\link{msigdb_h_table}},
\code{\link{pcnet_table}},
\code{\link{string_table}}
}
\concept{datasets}
\keyword{datasets}
