% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goa_bp_table_data.R
\docType{data}
\name{goa_bp_table}
\alias{goa_bp_table}
\title{GOA biological process annotation table}
\format{data.frame object with: Gene.symbol, GO.ID, GO.name}
\source{
"GOA Homo sapiens protein dataset, 2020-03-23 release. Downloaded on 2020-03-30 (http://current.geneontology.org/products/pages/downloads.html)."
}
\usage{
goa_bp_table
}
\description{
A table containing 140,067 GO biological process annotations (17,814 unique gene symbols, 12,362 unique GO BP terms).
}
\seealso{
Other datasets: 
\code{\link{gnomad_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{msigdb_table}}
}
\concept{datasets}
\keyword{datasets}
