% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goa_cc_table_data.R
\docType{data}
\name{goa_cc_table}
\alias{goa_cc_table}
\title{GOA cellular component annotation table}
\format{data.frame object with: Gene.symbol, GO.ID, GO.name}
\source{
"GOA Homo sapiens protein dataset, 2020-03-23 release. Downloaded on 2020-03-30 (http://current.geneontology.org/products/pages/downloads.html)."
}
\usage{
goa_cc_table
}
\description{
A table containing 83,029 GO cellular component annotations (18,880 unique gene symbols, 1,765 unique GO CC terms).
}
\seealso{
Other datasets: 
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_mf_table}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{msigdb_table}}
}
\concept{datasets}
\keyword{datasets}
