% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_table_data.R
\docType{data}
\name{gwas_table}
\alias{gwas_table}
\title{GWAS catalog table}
\format{data.frame object with: PUBMEDID, DISEASE.TRAIT, CHR_ID, CHR_POS, SNPS, P.VALUE, MAPPED_TRAIT, SUTDY.ACCESSION.}
\source{
"gwas_catalog_v1.0.2-associations_e98_r2020-03-08.tsv file downloaded from GWAS catalog: https://www.ebi.ac.uk/gwas/docs/file-downloads"
}
\usage{
gwas_table
}
\description{
A table containing 179,364 genetic associations in GWAS catalog. Genetic data were mapped to genome assembly GRCh38.p13 and dbSNP Build 152.
}
\keyword{datasets}
