% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gwas_table}
\alias{gwas_table}
\title{NHGRI-EBI GWAS catalog data}
\format{
A tibble object with 511,955 rows and 5 columns:
\describe{
  \item{PUBMEDID}{<int> PubMed identification number}
  \item{DISEASE.TRAIT}{<chr> Disease or trait examined in study}
  \item{SNP}{<chr> current rs number for strongest SNP (dbSNP build 154)}
  \item{P.VALUE}{<dbl> Reported p-value for strongest SNP}
  \item{STUDY.ACCESSION}{<chr> Accession ID allocated to a GWAS Catalog study}
}
}
\source{
All associations v1.0.2 dataset (2023-11-08 release) downloaded from:
\url{https://www.ebi.ac.uk/gwas/docs/file-downloads}
}
\usage{
gwas_table
}
\description{
Published trait-SNP associations in the NHGRI-EBI GWAS catalog.
}
\references{
Sollis E, Mosaku A, Abid A, et al. The NHGRI-EBI GWAS Catalog: knowledgebase and deposition resource.
Nucleic Acids Res. 2023;51(D1):D977-D985. doi:10.1093/nar/gkac1010
}
\seealso{
Other datasets: 
\code{\link{accession_gene_table}},
\code{\link{bioplex_table}},
\code{\link{example_data2}},
\code{\link{example_data3}},
\code{\link{example_data}},
\code{\link{genes_snps_hash}},
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gtex_brain_table}},
\code{\link{gtex_protein_table}},
\code{\link{gtex_rna_table}},
\code{\link{hgnc_group_table}},
\code{\link{hpa_rna_table}},
\code{\link{inweb_table}},
\code{\link{irefindex_table}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c2_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_c8_table}},
\code{\link{msigdb_h_table}},
\code{\link{string_table}}
}
\concept{datasets}
\keyword{datasets}
