% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gwas_table}
\alias{gwas_table}
\title{GWAS catalog table}
\format{data.frame object with: PUBMEDID, DISEASE.TRAIT, SNP, P.VALUE, SUTDY.ACCESSION}
\source{
gwas_catalog_v1.0.2-associations_e98_r2020-03-08.tsv file downloaded from GWAS catalog: https://www.ebi.ac.uk/gwas/docs/file-downloads.
}
\usage{
gwas_table
}
\description{
A table containing 190,668 genetic associations in GWAS catalog. SNP data are based on dbSNP Build 152.
}
\seealso{
Other datasets: 
\code{\link{accession_gene_table}},
\code{\link{example_data}},
\code{\link{genes_snps}},
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{hgnc_group_table}},
\code{\link{inweb_hash}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c2_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_h_table}}
}
\concept{datasets}
\keyword{datasets}
