% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgnc_group_table_data.R
\docType{data}
\name{hgnc_group_table}
\alias{hgnc_group_table}
\title{HGNC gene group annotation table}
\format{data.frame object with: Gene.symbol, Group.name}
\source{
"HGNC genes within groups dataset. Downloaded on 2020-03-30 (https://www.genenames.org/download/statistics-and-files/)."
}
\usage{
hgnc_group_table
}
\description{
A table containing 28,677 gene group annotations from HGNC (23,874 unique symbols, 1415 unique groups).
}
\seealso{
Other datasets: 
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gwas_table}},
\code{\link{msigdb_table}}
}
\concept{datasets}
\keyword{datasets}
