% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hpa_table}
\alias{hpa_table}
\title{Subsetted Human Protein Atlas annotation table}
\format{data.frame object with: tissue, gene, significant (indicating whether the gene is elvated in the tissue.)}
\source{
merge between proteinatlas.tsv.zip and normal_tissue.tsv.zip downloaded on 2020-06-12 (https://www.proteinatlas.org/about/download).
}
\usage{
hpa_table
}
\description{
A table containing classification of all human protein-coding genes based on transcript expression levels in 33 tissues
(19,634 unique genes). Only tissue with elevated expression has been included.
}
\note{
some tissue is divided into sub-categories, e.g. Small intestine (Intestine) and Small intestine (Intestine). See table 2-3 at \url{https://www.proteinatlas.org/humanproteome/tissue/tissue+specific} 
for additional details.
}
\seealso{
Other datasets: 
\code{\link{accession_gene_table}},
\code{\link{example_data2}},
\code{\link{example_data3}},
\code{\link{example_data}},
\code{\link{genes_snps}},
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{inweb_hash}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c2_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_h_table}}
}
\concept{datasets}
\keyword{datasets}
