% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inweb_hash_data.R
\docType{data}
\name{inweb_hash}
\alias{inweb_hash}
\title{InWeb hash table}
\format{hash object with each of 18,478 genes as key, vector of known InWeb interators as value}
\source{
"Genoppi-v3 InWeb_Combined_Oct2018.Rdata file created by April Kim"
}
\usage{
inweb_hash
}
\description{
A hash table containing InWeb interactions for 18,478 genes (generated October 2018).
}
\keyword{datasets}
