% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{irefindex_table}
\alias{irefindex_table}
\title{iRefIndex protein-protein interactions}
\format{
A tibble object with 776,608 rows and 3 columns:
\describe{
  \item{Gene1}{<chr> Gene name of first protein}
  \item{Gene2}{<chr> Gene name of second protein}
  \item{Score}{<int> Interaction confidence score (i.e. number of reporting PMIDs)}
}
}
\source{
iRefIndex 20.0 human MITAB file downloaded from:
\url{https://irefindex.vib.be/wiki/index.php/README_MITAB2.6_for_iRefIndex_20.0}
}
\usage{
irefindex_table
}
\description{
A dataset containing 776,608 undirected protein-protein interactions for
18,166 human genes in iRefIndex 20.0.
}
\references{
Razick S, Magklaras G, Donaldson IM. iRefIndex: a consolidated protein interaction database with provenance.
BMC Bioinformatics. 2008;9:405. Published 2008 Sep 30. doi:10.1186/1471-2105-9-405
}
\seealso{
Other datasets: 
\code{\link{accession_gene_table}},
\code{\link{bioplex_table}},
\code{\link{example_data2}},
\code{\link{example_data3}},
\code{\link{example_data}},
\code{\link{genes_snps_hash}},
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gtex_brain_table}},
\code{\link{gtex_protein_table}},
\code{\link{gtex_rna_table}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{hpa_rna_table}},
\code{\link{inweb_table}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c2_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_c8_table}},
\code{\link{msigdb_h_table}},
\code{\link{string_table}}
}
\concept{datasets}
\keyword{datasets}
