% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{irefindex_table}
\alias{irefindex_table}
\title{iRefIndex protein-protein interactions}
\format{
A data.frame with 396,984 rows and 5 variables:
\describe{
  \item{Gene1}{HGNC gene symbol of first protein}
  \item{Gene2}{HGNC gene symbol of second protein}
  \item{Score.hpr.max}{Highest PMID re-use}
  \item{Score.lpr.max}{Lowest PMID re-use}
  \item{Score.np.max}{Number of PMIDs}
}
}
\source{
iRefIndex data (v17.0) downloaded on 2021-01-06 from:
\url{https://irefindex.vib.be/download/irefindex/data/archive/release_17.0/psi_mitab/MITAB2.6/}
}
\usage{
irefindex_table
}
\description{
A dataset containing 396,984 published protein-protein interactions for 17,437 human genes
in iRefIndex 17.0.
}
\details{
This dataset was generated from the source BioPlex database by excluding all non-binary and non-human
interactions. HGNC gene symbols were then extracted from the database along with the relevant score columns. 
Rows corresponding to the same gene pairs were merged into a single row by keeping the maximum scores.
}
\references{
Razick S, Magklaras G, Donaldson IM. iRefIndex: a consolidated protein interaction database with provenance.
BMC Bioinformatics. 2008;9:405. Published 2008 Sep 30. doi:10.1186/1471-2105-9-405
}
\seealso{
Other datasets: 
\code{\link{accession_gene_table}},
\code{\link{bioplex_table}},
\code{\link{example_data2}},
\code{\link{example_data3}},
\code{\link{example_data}},
\code{\link{genes_snps}},
\code{\link{gnomad_table}},
\code{\link{goa_bp_table}},
\code{\link{goa_cc_table}},
\code{\link{goa_mf_table}},
\code{\link{gtex_protein}},
\code{\link{gtex_rna}},
\code{\link{gwas_table}},
\code{\link{hgnc_group_table}},
\code{\link{hpa_rna}},
\code{\link{inweb_table}},
\code{\link{msigdb_c1_table}},
\code{\link{msigdb_c2_table}},
\code{\link{msigdb_c3_table}},
\code{\link{msigdb_c4_table}},
\code{\link{msigdb_c5_table}},
\code{\link{msigdb_c6_table}},
\code{\link{msigdb_c7_table}},
\code{\link{msigdb_c8_table}},
\code{\link{msigdb_h_table}}
}
\concept{datasets}
\keyword{datasets}
