% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_interactive.R
\name{make_interactive}
\alias{make_interactive}
\title{Make ggplot interactive using plotly}
\usage{
make_interactive(p, source = NULL, legend = T, sig_text = "")
}
\arguments{
\item{legend}{boolean. Show legend for significant interactors?}

\item{x}{string. The x-column to be used from \code{p$overlay}.}

\item{y}{string. The y-column to be used from \code{p$overlay}.}
}
\description{
Takes a ggplot and turns it into 
an interactive plot.
}
\seealso{
Other interactivity: 
\code{\link{set_names_by_dataset}()}
}
\concept{interactivity}
