% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_gene_id.R
\name{map_gene_id}
\alias{map_gene_id}
\title{Map Uniprot accession ID to gene name (HGNC symbol)}
\usage{
map_gene_id(df)
}
\arguments{
\item{df}{data.frame containing accession_number column}
}
\value{
list of (df with accession_number column replaced by gene column, df with accession_number and gene columns)
}
\description{
Map and replace accession_number column in a data.frame with gene column. Isoforms (indicated by suffixes separated by . or - in acession_number) are mapped to the same gene names.
}
