% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_gene_id.R
\name{map_gene_id}
\alias{map_gene_id}
\title{map id to gene}
\usage{
map_gene_id(uniprot)
}
\arguments{
\item{uniprot}{a vector of uniprot IDs without isoform.}
}
\value{
a vector of corresponding hgnc symbols
}
\description{
maps an uniprot id to hgnc symbol using a hashmap
}
