% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_gene_id.R
\name{map_gene_id}
\alias{map_gene_id}
\title{map id to gene}
\usage{
map_gene_id(df)
}
\arguments{
\item{df}{data.frame containing accession_number column}
}
\value{
list of (df with accession_number column replaced by gene column, df with accession_number and gene columns)
}
\description{
maps an uniprot id to hgnc symbol using a hashmap
}
