% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_gene_id.R
\name{map_gene_id}
\alias{map_gene_id}
\title{Map Uniprot accession ID to gene name (HGNC symbol)}
\usage{
map_gene_id(df)
}
\arguments{
\item{df}{data.frame containing accession_number column}
}
\value{
data.frame consisting of input df + new gene column
}
\description{
Add gene column to data.frame containing accesion_number column. Isoforms (indicated by suffixes separated by . or - in acession_number) are mapped to the same gene names.
}
