% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overlay.R
\name{plot_overlay}
\alias{plot_overlay}
\title{plot overlay}
\usage{
plot_overlay(p, reference, point_expansion = 1.1)
}
\arguments{
\item{p}{a plot object returned from ggplot.}

\item{reference}{a list of named data.frames.}

\item{point_expansion}{percentage expansion of genelist points.}
}
\description{
Overlays a ggplot with a reference data set, e.g. a bait or a genelist
}
\examples{
\dontrun{
## generate a random plot
set.seed(3)
df = data.frame(gene=letters, fdr=runif(26), pvalue=runif(26), logFC = rnorm(26), significant = c(rep(T,10), rep(F, 16)))
p = plot_volcano(df) + labs(title='Random generated data and gglabs')

## Generate random dataset
ref1= data.frame(gene=c('b'),col_significant='brown',col_other='black')
ref2= data.frame(gene=c('d','e','f'),col_significant='blue',col_other='grey')
reference = list(ref1, ref2)
names(reference) = c('ASD genes', 'SCZ genes')

## 
p1 = plot_overlay(p, reference)
p1
p2 = plot_overlay(p1, list(bait=data.frame(gene='x', col_significant = 'red', col_other = 'orange')))
p2
}
}
