% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter_basic.R
\name{plot_scatter_basic}
\alias{plot_scatter_basic}
\title{Draw a scatter plot of replicates}
\usage{
plot_scatter_basic(
  df,
  bait = NULL,
  size_point = 3,
  col_signficant = "#41AB5D",
  col_other = "grey"
)
}
\arguments{
\item{df}{A data.frame containing at least gene, significant, replicate/triplicate columns.}

\item{bait}{the name of the bait}

\item{size_point}{size of point.}

\item{size_text}{size of the text label.}
}
\value{
a list of ggplot objects and replicate correlations
}
\description{
Draws replicate correlation scatter plot(s)
}
\examples{
\dontrun{
df = data.frame(gene = letters, rep1=rnorm(26), rep2=rnorm(26), rep3=rnorm(26), significant = T)
dat = plot_scatter_basic(df, bait = 'A')
dat

df = data.frame(gene = letters, rep1=rnorm(26), rep2=rnorm(26), significant = T)
dat = plot_scatter_basic(df, bait = 'A')
dat
 
}
}
