% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter_basic.R
\name{plot_scatter_basic}
\alias{plot_scatter_basic}
\title{plot a basic scatte plot}
\usage{
plot_scatter_basic(
  df,
  repA = "rep1",
  repB = "rep2",
  size_point = 3,
  col_significant = "#41AB5D",
  col_other = "grey"
)
}
\arguments{
\item{df}{A data.frame containing at least gene, significant, replicate/triplicate columns.}

\item{repA}{string that is column in df. Expects name to be in the form of rep[0-9], e.g. 'rep1'.}

\item{repB}{string that is column in df.}

\item{size_point}{size of point.}

\item{col_significant}{color for significant protein interactors.}

\item{col_other}{color for other protein interactors.}
}
\value{
a gg scatter plot.
}
\description{
Draws a pair of specific replicates in a scatter plot.
}
