% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano_basic.R
\name{plot_volcano_basic}
\alias{plot_volcano_basic}
\title{Plot basic volcano}
\usage{
plot_volcano_basic(
  df,
  col_significant = "#41AB5D",
  col_other = "grey",
  gg.size = 3
)
}
\arguments{
\item{df}{a data.frame with at least columns gene, logFC, pvalue and significant.}

\item{col_significant}{the color of significant proteins/rows.}

\item{col_other}{the color of non-significnt proteins/rows.}

\item{gg.size}{the size of the points.}
}
\description{
takes in a data.frane with the columns gene, logFC, pvalue and significant 
to draw a volcano. Optionally, a column indicating 'color' (string) can be supplied to 
indicate the volcano color scheme and whether to draw names of specific proteins.
}
