% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_read.R
\name{read_input}
\alias{read_input}
\title{read input}
\usage{
read_input(filename, header, sep)
}
\arguments{
\item{filename}{path to input file}

\item{header}{a logical value indicating whether the file contains the names of the variables as its first line.}

\item{sep}{the field character seperator, see \code{?read.table}.}
}
\description{
reads
}
