% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_input.R
\name{read_input}
\alias{read_input}
\title{Read input data}
\usage{
read_input(filename, header = "auto", sep = "auto")
}
\arguments{
\item{filename}{path to input file}

\item{header}{a logical value indicating whether the file contains the names of the variables as its first line}

\item{sep}{the field character seperator, see \code{?fread}}
}
\value{
list contaiing input data.table and list of boolean variables indicating input format
}
\description{
Read in input file and store as data.frame. Call check_input() to check data format.
}
