% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{shape_to_symbol}
\alias{shape_to_symbol}
\title{translate ggplot shapes to plotly symbols}
\usage{
shape_to_symbol(vec)
}
\arguments{
\item{vec}{a character vector of ggplot shapes.}
}
\description{
translate ggplot shapes to plotly symbols
}
\seealso{
Other misc: 
\code{\link{\%nin\%}()},
\code{\link{as.bait}()},
\code{\link{bold}()},
\code{\link{catf}()},
\code{\link{color_distinct}()},
\code{\link{color_gradient}()},
\code{\link{enumerate_replicate_combinations}()},
\code{\link{italics}()},
\code{\link{lun}()},
\code{\link{null_omit}()},
\code{\link{symbol_to_shape}()}
}
\concept{misc}
