% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_invalid.R
\name{stop_invalid_columns}
\alias{stop_invalid_columns}
\title{Stop invalid columns}
\usage{
stop_invalid_columns(data, prefix, need)
}
\arguments{
\item{data}{a data.frame}

\item{prefix}{e.g. 'snpData'}

\item{need}{vector of strings needed in data}
}
\description{
yields an informative error message
if the a data object doesnt contain the correct
or verified column names.
}
