% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{table_symbols}
\alias{table_symbols}
\title{get table of plotly/ggplot symbols}
\usage{
table_symbols()
}
\description{
get a table for translating
between plotly symbols and ggplot shapes.
This is primiarly used in the shiny application
to ensure consistency between ggplot and plotly.
}
\seealso{
Other misc: 
\code{\link{\%nin\%}()},
\code{\link{as.bait}()},
\code{\link{as.goi}()},
\code{\link{bold}()},
\code{\link{catf}()},
\code{\link{color_distinct}()},
\code{\link{color_gradient}()},
\code{\link{enumerate_replicate_combinations}()},
\code{\link{get_gg_legend}()},
\code{\link{hline}()},
\code{\link{italics}()},
\code{\link{lun}()},
\code{\link{null_omit}()},
\code{\link{shape_to_symbol}()},
\code{\link{symbol_to_shape}()},
\code{\link{vline}()}
}
\concept{misc}
