% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overlay.R
\name{validate_reference}
\alias{validate_reference}
\title{validate reference data.frame}
\usage{
validate_reference(
  df,
  valid = c("gene", "col_significant", "col_other", "shape", "dataset", "stroke",
    "alt_label", "label", "size", "symbol", "pLI", "opacity", "size")
)
}
\arguments{
\item{df}{a data.frame}

\item{valid}{a vector of valid ggplot options}
}
\value{
a data.frame
}
\description{
A function that checks the column names 
of a data.frame to see whether they contain values
that can be used by ggplot.
}
\note{
internal
}
